### 2024-04-30 bc01cf7
Update paper.bib
make DOIs valid
