+++ paper.bib
@@ -6,7 +6,7 @@ number = {1},
 pages = {1-29},
 year = {2008},
 issn = {0079-6611},
-doi = {https://doi.org/10.1016/j.pocean.2008.02.002},
+doi = {10.1016/j.pocean.2008.02.002},
 url = {https://www.sciencedirect.com/science/article/pii/S0079661108000232},
 author = {J.H. LaCasce}
 }
@@ -18,7 +18,7 @@ volume = {121},
 pages = {49-75},
 year = {2018},
 issn = {1463-5003},
-doi = {https://doi.org/10.1016/j.ocemod.2017.11.008},
+doi = {10.1016/j.ocemod.2017.11.008},
 url = {https://www.sciencedirect.com/science/article/pii/S1463500317301853},
 author = {Erik {van Sebille} and Stephen M. Griffies and Ryan Abernathey and Thomas P. Adams and Pavel Berloff and Arne Biastoch and Bruno Blanke and Eric P. Chassignet and Yu Cheng and Colin J. Cotter and Eric Deleersnijder and Kristofer Döös and Henri F. Drake and Sybren Drijfhout and Stefan F. Gary and Arnold W. Heemink and Joakim Kjellsson and Inga Monika Koszalka and Michael Lange and Camille Lique and Graeme A. MacGilchrist and Robert Marsh and C. Gabriela {Mayorga Adame} and Ronan McAdam and Francesco Nencioli and Claire B. Paris and Matthew D. Piggott and Jeff A. Polton and Siren Rühs and Syed H.A.M. Shah and Matthew D. Thomas and Jinbo Wang and Phillip J. Wolfram and Laure Zanna and Jan D. Zika}
 }
