### 2024-04-30 #6705 @selipot
@kthyng Thank you. Are you referring to lines 47 and 59 where the text goes beyond the text width?
