### 2024-04-30 dceba96
Update paper.md
correct hyperlink to AWS open data page
