### 2024-04-30 564a539
Update paper.md
try fixing figure reference/label and text overspilling
