### 2024-04-30 019f464
Update paper.md
try to fix figure reference/label
