### 2024-04-30 #6705 @selipot
> Those but also there is a reference ? in the I think it is a figure caption and a markdown link isn't working I think just due to typo.

- I was able to fix the markdown link by correcting the typo.

- I managed to fix the text overflow on line 59 but I am unsure how to force a carriage return for line 47. Any advice?

- I am unable to make the internal reference work for the figure despite following the template. I just tried to change the order of `\label` and `\autoref` but that does not fix it. Also the figure caption does not show. Not sure how to fix this.
