+++ paper.md
@@ -54,11 +54,15 @@ In Earth, Ocean, Geo-, and Atmospheric Science, *Eulerian data* typically refers
 
 The scope of the Clouddrift library includes: 
 
-1. **Working with contiguous ragged array representations of data, whether they originate from geosciences or any other field**. Ragged array representations are useful when the data lengths of the instances of a feature (variable) are not all equal. With such representations the data for each feature are stored contiguously in memory, and the number of elements that each feature has is contained in a count variable which Clouddrift calls *rowsize*. A graphical representation of the application of the ragged array structure to Lagrangian data is displayed in \autoref{fig:raggedarray}.
+1. **Working with contiguous ragged array representations of data, whether they originate from geosciences or any other field**.
+
+Ragged array representations are useful when the data lengths of the instances of a feature (variable) are not all equal. With such representations the data for each feature are stored contiguously in memory, and the number of elements that each feature has is contained in a count variable which Clouddrift calls *rowsize*. A graphical representation of the application of the ragged array structure to Lagrangian data is displayed in \autoref{fig:raggedarray}.
 
 ![Ragged array representation for Lagrangian data. \label{fig:raggedarray}](ragged_array.png){ width=95% }   
 
-2. **Delivering functions and methods to perform scientific analysis of Lagrangian data, oceanographic or otherwise [@Lacasce:2008;@Vansebille:2018], structured as ragged arrays or otherwise**. A straightforward example of Lagrangian analysis provided by Clouddrift is the derivation of Lagrangian velocities from a sequence of Lagrangian positions, and vice versa. Another more involved example is the discovery of pairs of Lagrangian data prescribed by distances in space and time. Both of these methods are currently available with Clouddrift.
+2. **Delivering functions and methods to perform scientific analysis of Lagrangian data, oceanographic or otherwise [@Lacasce:2008;@Vansebille:2018], structured as ragged arrays or otherwise**.
+
+A straightforward example of Lagrangian analysis provided by Clouddrift is the derivation of Lagrangian velocities from a sequence of Lagrangian positions, and vice versa. Another more involved example is the discovery of pairs of Lagrangian data prescribed by distances in space and time. Both of these methods are currently available with Clouddrift.
 
 *Example:* The following example combines the use of the clouddrift function 
 `velocity_from_position` from the `clouddrift.kinematics` module, with the function `apply_ragged` from the `clouddrift.ragged` module, in order to calculate the velocities of multiple particles, the coordinates of which are found in the ragged arrays `x`, `y`, and `t` that share row sizes 2, 3, and 4:  
@@ -75,7 +79,9 @@ u1, v1 = apply_ragged(velocity_from_position, [x, y, t], rowsize,
   coord_system="cartesian")
 ```
 
-3. **Processing publicly available Lagrangian datasets into the common ragged array data structure and format**. Through data *adapters*, this type of processing includes not only converting Lagrangian data from typically regular arrays to ragged arrays but also aggregating data and metadata from multiple data files into a single data file. The canonical example of the Clouddrift library is constituted of the data from the NOAA Global Drifter Program [@Elipot:2022].
+3. **Processing publicly available Lagrangian datasets into the common ragged array data structure and format**.
+
+Through data *adapters*, this type of processing includes not only converting Lagrangian data from typically regular arrays to ragged arrays but also aggregating data and metadata from multiple data files into a single data file. The canonical example of the Clouddrift library is constituted of the data from the NOAA Global Drifter Program [@Elipot:2022].
 
[…]
