+++ paper.md
@@ -60,9 +60,7 @@ Ragged array representations are useful when the data lengths of the instances o
 
 ![Ragged array representation for Lagrangian data. \label{fig:raggedarray}](ragged_array.png){ width=95% }   
 
-2. **Delivering functions and methods to perform scientific analysis of Lagrangian data, oceanographic or otherwise [@Lacasce:2008;@Vansebille:2018], structured as ragged arrays or otherwise**.
-
-A straightforward example of Lagrangian analysis provided by Clouddrift is the derivation of Lagrangian velocities from a sequence of Lagrangian positions, and vice versa. Another more involved example is the discovery of pairs of Lagrangian data prescribed by distances in space and time. Both of these methods are currently available with Clouddrift.
+2. **Delivering functions and methods to perform scientific analysis of Lagrangian data, oceanographic or otherwise [@Lacasce:2008;@Vansebille:2018], structured as ragged arrays or otherwise**. A straightforward example of Lagrangian analysis provided by Clouddrift is the derivation of Lagrangian velocities from a sequence of Lagrangian positions, and vice versa. Another more involved example is the discovery of pairs of Lagrangian data prescribed by distances in space and time. Both of these methods are currently available with Clouddrift.
 
 *Example:* The following example combines the use of the clouddrift function 
 `velocity_from_position` from the `clouddrift.kinematics` module, with the function `apply_ragged` from the `clouddrift.ragged` module, in order to calculate the velocities of multiple particles, the coordinates of which are found in the ragged arrays `x`, `y`, and `t` that share row sizes 2, 3, and 4:  
@@ -79,9 +77,7 @@ u1, v1 = apply_ragged(velocity_from_position, [x, y, t], rowsize,
   coord_system="cartesian")
 ```
 
-3. **Processing publicly available Lagrangian datasets into the common ragged array data structure and format**.
-
-Through data *adapters*, this type of processing includes not only converting Lagrangian data from typically regular arrays to ragged arrays but also aggregating data and metadata from multiple data files into a single data file. The canonical example of the Clouddrift library is constituted of the data from the NOAA Global Drifter Program [@Elipot:2022].
+3. **Processing publicly available Lagrangian datasets into the common ragged array data structure and format**. Through data *adapters*, this type of processing includes not only converting Lagrangian data from typically regular arrays to ragged arrays but also aggregating data and metadata from multiple data files into a single data file. The canonical example of the Clouddrift library is constituted of the data from the NOAA Global Drifter Program [@Elipot:2022].
 
 *Example:* The following example locally builds an xarray dataset, with ragged array representations, of the latest dataset of position, velocity, and sea surface temperature from the Global Drifter Program quality-controlled 6-hour interpolated data from ocean surface drifting buoys:
 ```
@@ -89,9 +85,7 @@ from clouddrift.adapters import gdp6h
 ds = gdp6h.to_raggedarray().to_xarray()
 ```
 
-4. **Making cloud-optimized ragged array datasets easily accessible**.
-
-This involves opening in a computing environment, without unnecessary download, Lagrangian datasets available from cloud servers, as well as opening Lagrangian dataset which have been seamlessly processed by the Clouddrift data *adapters*.    
+4. **Making cloud-optimized ragged array datasets easily accessible**. This involves opening in a computing environment, without unnecessary download, Lagrangian datasets available from cloud servers, as well as opening Lagrangian dataset which have been seamlessly processed by the Clouddrift data *adapters*.    
 
 *Example:* The following simple command remotely opens without download the hourly location, current velocity, and temperature collected from Global Drifter Program drifters world-wide, distributed as a zarr archive with ragged array representations and stored in a cloud storage as part of the [Registry of Open Data on AWS](https://registry.opendata.aws/noaa-oar-hourly-gdp/):
 
