+++ .github/workflows/draft-pdf.yml
@@ -20,7 +20,7 @@ jobs:
           journal: joss
           paper-path: paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: paper.pdf
