+++ .github/workflows/draft-pdf.yml
@@ -12,8 +12,6 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-        with:
-          ref: paper
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
