### 2024-05-01 f263044
simplify adapter
