+++ clouddrift/adapters/gdp/rawfiles.py
@@ -1,16 +1,15 @@
 from __future__ import annotations
 
 import datetime
-import gzip
 import logging
-import math
 import os
 import tempfile
-from abc import abstractmethod
-from collections import defaultdict
-from dataclasses import dataclass, field
-from typing import Literal, override
+from dataclasses import dataclass
+from typing import Callable, Literal
 
+import numpy as np
+import pandas as pd
+import xarray as xr
 from tqdm import tqdm
 
 from clouddrift.adapters.gdp import get_gdp_metadata
@@ -30,194 +29,113 @@ _RecordKind = Literal["position"] | Literal["sensor"] | Literal["raw"]
 _logger = logging.getLogger(__name__)
 
 
-@dataclass
-class DrifterMetadata:
-    id: int = field(
-        metadata=dict()
-    )
-    wmoNumber: int = field(
-        metadata=dict()
-    )
-    programNumber: int = field(
-        metadata=dict()
-    )
-    buoyType: str = field(
-        metadata=dict()
-    )
-    startDateTime: datetime.datetime = field(
-        metadata=dict()
-    )
-    startLat: float = field(
-        metadata=dict()
-    )
-    startLon: float = field(
-        metadata=dict()
-    )
-    endDateTime: datetime.datetime = field(
-        metadata=dict()
-    )
-    endLat: float = field(
-        metadata=dict()
-    )
-    endLon: float = field(
-        metadata=dict()
-    )
-    drogueOffDateTime: str = field(
-        metadata=dict()
-    )
-    deathCode: int = field(
-        metadata=dict()
-    )
+def _parse_datetime_with_day_ratio(monthSeries: np.ndarray, day_series: np.ndarray, year_series: np.ndarray):
+    values = list()
+    for month, day_with_ratio, year in zip(monthSeries, day_series, year_series):
+        day = day_with_ratio // 1
+        dayratio = day_with_ratio - day
+        seconds = dayratio * _SECONDS_IN_DAY
+        dt_ns = (datetime.datetime(year=int(year), month=int(month), day=int(day)) + datetime.timedelta(seconds=seconds)).timestamp() * 10**9
+        values.append(int(dt_ns))
+    return np.array(values, dtype="datetime64[ns]")
 
 
 @dataclass
-class RawDrifterBase:
-    id: int = field(
-        metadata=dict()
-    )
+class ParsingConfiguration:
+    cols: list[str]
+    coords: list[str]
+    col_dtypes: dict[str, np.dtype]
+    transform: dict[str, tuple[list[str], Callable[..., np.ndarray]]]
 
-    @classmethod
-    @abstractmethod
-    def fromcols(cls, cols: list[str]) -> RawDrifterBase:
-        raise NotImplementedError()
+    def get_vars_config_map(self, dim_name: str, df: pd.DataFrame):
+        all_config_map = self._get_all_config_map(dim_name, df)
+        return {col: all_config_map[col] for col in all_config_map.keys() if col not in self.coords}
 
+    def get_coords_config_map(self, dim_name: str, df: pd.DataFrame):
+        all_config_map = self._get_all_config_map(dim_name, df)
+        return {col: all_config_map[col] for col in self.coords}
 
-@dataclass
-class RawDrifterPosition:
-    positionDateTime: datetime.datetime = field(
-        metadata=dict()
-    )
-    lat: float = field(
-        metadata=dict()
-    )
-    lon: float = field(
-        metadata=dict()
-    )
-    qualityIndex: float
+    def _get_all_config_map(self, dim_name: str, df: pd.DataFrame):
+        pre_transform = {col: ([dim_name], df[[col]].to_numpy().flatten()) for col in self.cols}
+        post_transform = self._apply_transform(pre_transform)
+        return post_transform
[…]
