+++ clouddrift/adapters/gdp/rawfiles.py
@@ -19,9 +19,6 @@ from clouddrift.raggedarray import RaggedArray
 _DATA_URL = "https://www.aoml.noaa.gov/ftp/pub/phod/pub/pazos/data/shane/sst"
 _TMP_PATH = os.path.join(tempfile.gettempdir(), "clouddrift", "gdpraw")
 _FILNAME_TEMPLATE = "buoydata_{start}_{end}_{suffix}.dat.gz"
-os.makedirs(
-    _TMP_PATH, exist_ok=True
-)  # generate temp directory for hurdat related intermerdiary data
 _SECONDS_IN_DAY = 86_400
 
 _RecordKind = Literal["position"] | Literal["sensor"] | Literal["raw"]
@@ -29,13 +26,18 @@ _RecordKind = Literal["position"] | Literal["sensor"] | Literal["raw"]
 _logger = logging.getLogger(__name__)
 
 
-def _parse_datetime_with_day_ratio(monthSeries: np.ndarray, day_series: np.ndarray, year_series: np.ndarray):
+def _parse_datetime_with_day_ratio(
+    monthSeries: np.ndarray, day_series: np.ndarray, year_series: np.ndarray
+):
     values = list()
     for month, day_with_ratio, year in zip(monthSeries, day_series, year_series):
         day = day_with_ratio // 1
         dayratio = day_with_ratio - day
         seconds = dayratio * _SECONDS_IN_DAY
-        dt_ns = (datetime.datetime(year=int(year), month=int(month), day=int(day)) + datetime.timedelta(seconds=seconds)).timestamp() * 10**9
+        dt_ns = (
+            datetime.datetime(year=int(year), month=int(month), day=int(1))
+            + datetime.timedelta(days=int(day), seconds=seconds)
+        ).timestamp() * 10**9
         values.append(int(dt_ns))
     return np.array(values, dtype="datetime64[ns]")
 
@@ -45,24 +47,45 @@ class ParsingConfiguration:
     cols: list[str]
     coords: list[str]
     col_dtypes: dict[str, np.dtype]
+    remove: list[Callable[[pd.DataFrame], pd.DataFrame]]
     transform: dict[str, tuple[list[str], Callable[..., np.ndarray]]]
 
     def get_vars_config_map(self, dim_name: str, df: pd.DataFrame):
         all_config_map = self._get_all_config_map(dim_name, df)
-        return {col: all_config_map[col] for col in all_config_map.keys() if col not in self.coords}
+        return {
+            col: all_config_map[col]
+            for col in all_config_map.keys()
+            if col not in self.coords
+        }
 
     def get_coords_config_map(self, dim_name: str, df: pd.DataFrame):
         all_config_map = self._get_all_config_map(dim_name, df)
         return {col: all_config_map[col] for col in self.coords}
 
     def _get_all_config_map(self, dim_name: str, df: pd.DataFrame):
-        pre_transform = {col: ([dim_name], df[[col]].to_numpy().flatten()) for col in self.cols}
+        post_remove = self._apply_remove(df)
+        pre_transform = dict()
+
+        for col in self.cols:
+            data_array = (
+                post_remove[[col]].to_numpy().flatten().astype(self.col_dtypes[col])
+            )
+            pre_transform[col] = ([dim_name], data_array)
         post_transform = self._apply_transform(pre_transform)
         return post_transform
 
-    def _apply_transform(self, variable_config_map: dict[str, tuple[list[str], pd.DataFrame]]):
+    def _apply_remove(self, df: pd.DataFrame):
+        temp_df = df
+        for filter_ in self.remove:
+            mask = filter_(temp_df)
+            temp_df = temp_df[~mask]
+        return temp_df
+
+    def _apply_transform(
[…]
