### 2024-05-07 bb7705e
simplify and add gdpraw entry
