### 2024-05-07 5bafb06
simplify and add gdpraw entry
