+++ clouddrift/adapters/gdp/rawfiles.py
@@ -95,6 +95,19 @@ class ParsingConfiguration:
         return variable_config_map
 
 
+def _future_years_mask(df):
+    return df["obsYear"] > datetime.datetime.now().year
+
+def _future_years_mask_sen(df):
+    return df["posObsYear"] > datetime.datetime.now().year
+
+def _future_years_mask_pos(df):
+    return df["senObsYear"] > datetime.datetime.now().year
+
+def _bad_drogue_values_mask(df):
+    return df["drogue"].astype(np.str_).str.match(r"[\.]{2,}")
+
+
 def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
     return {
         "position": ParsingConfiguration(
@@ -116,8 +129,8 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
                 )
             },
             remove=[
-                lambda df: df["obsYear"] > datetime.datetime.now().year,
-                lambda df: df["drogue"].str.match(r"[\.]{2,}")
+                _future_years_mask,
+                _bad_drogue_values_mask
             ],
             coords=["id", "obsDatetime"],
         ),
@@ -154,8 +167,8 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
                 )
             },
             remove=[
-                lambda df: df["obsYear"] > datetime.datetime.now().year,
-                lambda df: df["drogue"].str.match(r"[\.]{2,}")
+                _future_years_mask,
+                _bad_drogue_values_mask
             ],
             coords=["id", "obsDatetime"],
         ),
@@ -209,9 +222,9 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
                 ),
             },
             remove=[
-                lambda df: df["posObsYear"] > datetime.datetime.now().year,
-                lambda df: df["senObsYear"] > datetime.datetime.now().year,
-                lambda df: df["drogue"].astype(np.str_).str.match(r"([\d]*\.){2,}")
+                _future_years_mask_pos,
+                _future_years_mask_sen,
+                _bad_drogue_values_mask
             ],
             coords=["id", "posObsDatetime", "sensorObsDatetime"],
         ),
@@ -234,83 +247,84 @@ def _get_download_list(kind: _RecordKind) -> list[tuple[str, str]]:
     return requests
 
 
-def setup_rowsize(df: pd.DataFrame, config: ParsingConfiguration):
-    def wrapped_rowsize(id_) -> int:
-        traj_data_df = df[df["id"] == id_[0]]
-        coords = config.get_coords_config_map("obs", traj_data_df)
-        _, var = coords[
-            list(coords.keys())[0]
-        ]  # any of the coords will work, to determine the rowsize
-        return len(var)
+def rowsize(id_, **kwargs) -> int:
+    df: pd.DataFrame = kwargs.get("data_df")
+    config: ParsingConfiguration = kwargs.get("config")
 
-    return wrapped_rowsize
+    traj_data_df = df[df["id"] == id_[0]]
+    coords = config.get_coords_config_map("obs", traj_data_df)
+    _, var = coords[
+        list(coords.keys())[0]
+    ]  # any of the coords will work, to determine the rowsize
+    return len(var)
 
 
-def setup_preprocess(
-    md_df: pd.DataFrame, data_df: pd.DataFrame, config: ParsingConfiguration
-):
-    def wrapped_preprocess(id_) -> xr.Dataset:
-        traj_md_df = md_df[md_df["ID"] == id_[0]]
-        traj_data_df = data_df[data_df["id"] == id_[0]]
+def preprocess(id_, **kwargs) -> xr.Dataset:
+    md_df: pd.DataFrame = kwargs.get("md_df")
[…]
