+++ clouddrift/raggedarray.py
@@ -6,8 +6,17 @@ Datasets and Awkward Arrays.
 
 from __future__ import annotations
 
+import asyncio
+import logging
+
+# import multiprocessing.dummy as mp
+import math
+import multiprocessing as mp
+import os
+import typing
 import warnings
 from collections.abc import Callable
+from itertools import batched
 from typing import Any, Literal
 
 import awkward as ak  # type: ignore
@@ -19,6 +28,18 @@ from clouddrift.ragged import rowsize_to_index
 
 DimNames = Literal["rows", "obs"]
 _DISABLE_SHOW_PROGRESS = False  # purely to de-noise our test suite output, should never be used/configured outside of that.
+_logger = logging.getLogger(__name__)
+
+_NamedNdArrayDict = dict[str, np.ndarray]
+_JobResult = typing.NamedTuple(
+    "_JobResult",
+    [
+        ("coords", _NamedNdArrayDict),
+        ("metadata", _NamedNdArrayDict),
+        ("data", _NamedNdArrayDict),
+        ("coord_dims", dict[str, str]),
+    ],
+)
 
 
 class RaggedArray:
@@ -108,6 +129,7 @@ class RaggedArray:
         rowsize_func: Callable[[int], int] | None = None,
         attrs_global: dict | None = None,
         attrs_variables: dict | None = None,
+        parallelized: bool = False,
         **kwargs,
     ):
         """Generate a ragged array archive from a list of files
@@ -138,17 +160,32 @@ class RaggedArray:
             if rowsize_func
             else lambda i, **kwargs: preprocess_func(i, **kwargs).sizes["obs"]
         )
-        rowsize = cls.number_of_observations(rowsize_func, indices, **kwargs)
-        coords, metadata, data, coord_dims = cls.allocate(
-            preprocess_func,
-            indices,
-            rowsize,
-            name_coords,
-            name_meta,
-            name_data,
-            name_dims,
-            **kwargs,
-        )
+
+        if parallelized:
+            async def parallel_exec():
+                return await cls._get_vars_parallel(
+                    indices,
+                    preprocess_func,
+                    name_coords,
+                    name_meta,
+                    name_data,
+                    name_dims,
+                    rowsize_func,
+                    **kwargs,
+                )
+            results = asyncio.run(parallel_exec())
+            coords, metadata, data, coord_dims = results
+        else:
+            coords, metadata, data, coord_dims = cls._get_vars_serial(
+                indices,
+                preprocess_func,
+                name_coords,
+                name_meta,
+                name_data,
+                name_dims,
+                rowsize_func,
+                **kwargs,
+            )
 
         extracted_attrs_global, extracted_attrs_variables = cls.attributes(
             preprocess_func(indices[0], **kwargs),
@@ -166,6 +203,157 @@ class RaggedArray:
         return RaggedArray(
             coords, metadata, data, attrs_global, attrs_variables, name_dims, coord_dims
         )
+
+    @classmethod
+    def _get_vars_serial(
+        cls,
+        indices: list[int],
+        preprocess_func: Callable[[int], xr.Dataset],
+        name_coords: list,
+        name_meta: list = list(),
+        name_data: list = list(),
+        name_dims: dict[str, DimNames] = {},
+        rowsize_func: Callable[[int], int] | None = None,
+        chunk_size: int = 5000,
+        **kwargs,
+    ):
+        coords: _NamedNdArrayDict = dict()
+        metadata: _NamedNdArrayDict = dict()
+        data: _NamedNdArrayDict = dict()
+        coord_dims: dict[str, str] | None = None
+
+        for ids in tqdm(
+            batched(indices, chunk_size),
+            desc="processing chunks",
[…]
