### 2024-05-08 092182e
🔍 update readme, badges, social connects  (#431)
* 🔍 update readme

* edit README.md

* Add twitter account

---------

Co-authored-by: Shane Elipot <selipot@miami.edu>
