### 2024-05-09 #6705 @AnjaliSandip
👋 @malmans2, @dionhaefner, @SaumikDana & @felixcremer, would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
