### 2024-05-09 537291e
update dev install instructions
