### 2024-05-09 c2693dc
update regex and remove parallelization stuff
