### 2024-05-10 f6d9737
chunk data files and process them into zarr archives to be combined into an aggergate final archive
