+++ clouddrift/adapters/gdp/rawfiles.py
@@ -1,10 +1,13 @@
 from __future__ import annotations
 
+import asyncio
 import datetime
 import logging
+import multiprocessing as mp
 import os
 import tempfile
 from dataclasses import dataclass
+from multiprocessing.pool import AsyncResult
 from typing import Callable, Literal
 
 import numpy as np
@@ -231,7 +234,7 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
     }.get(kind)
 
 
-def _get_download_list(kind: _RecordKind) -> list[tuple[str, str]]:
+def _get_download_list(tmp_path: str, kind: _RecordKind) -> list[tuple[str, str]]:
     suffix = {
         "position": "edited_pfiles",
         "sensor": "edited_sfiles",
@@ -243,7 +246,7 @@ def _get_download_list(kind: _RecordKind) -> list[tuple[str, str]]:
 
     for start, end in batches:
         filename = _FILNAME_TEMPLATE.format(start=start, end=end, suffix=suffix)
-        requests.append((f"{_DATA_URL}/{filename}", os.path.join(_TMP_PATH, filename)))
+        requests.append((f"{_DATA_URL}/{filename}", os.path.join(tmp_path, filename)))
     return requests
 
 
@@ -251,7 +254,7 @@ def rowsize(id_, **kwargs) -> int:
     df: pd.DataFrame = kwargs.get("data_df")
     config: ParsingConfiguration = kwargs.get("config")
 
-    traj_data_df = df[df["id"] == id_[0]]
+    traj_data_df = df[df["id"] == id_]
     coords = config.get_coords_config_map("obs", traj_data_df)
     _, var = coords[
         list(coords.keys())[0]
@@ -264,8 +267,8 @@ def preprocess(id_, **kwargs) -> xr.Dataset:
     data_df: pd.DataFrame = kwargs.get("data_df")
     config: ParsingConfiguration = kwargs.get("config")
 
-    traj_md_df = md_df[md_df["ID"] == id_[0]]
-    traj_data_df = data_df[data_df["id"] == id_[0]]
+    traj_md_df = md_df[md_df["ID"] == id_]
+    traj_data_df = data_df[data_df["id"] == id_]
 
     variables = {
         "wmo_number": (
@@ -320,48 +323,10 @@ def preprocess(id_, **kwargs) -> xr.Dataset:
     dataset = xr.Dataset(variables, coords=coords)
     return dataset
 
-
-def to_raggedarray(
-    kind: _RecordKind = "both",
-    tmp_path: str = _TMP_PATH,
-    max: int | None = None,
-) -> RaggedArray:
-    config = _get_parsing_config(kind)
-    requests = _get_download_list(kind)
-    destinations = [dst for (_, dst) in requests]
-    os.makedirs(tmp_path, exist_ok=True)
-
-    if max:
-        requests = requests[:max]
-        destinations = destinations[:max]
-
-    download_with_progress(requests)
-
-    df: pd.DataFrame = None
-    for fp in tqdm(destinations, desc="loading data files"):
-        before = datetime.datetime.now()
-
-        cur_df = pd.read_csv(
-            fp,
-            sep=r"\s+",
-            header=None,
-            names=config.cols,
-            engine="c",
-            compression="gzip",
-        )
-
-        after = datetime.datetime.now()
-
-        _logger.debug(f"elapsed time to load {len(cur_df)} records :: {after - before}")
-
-        if df is None:
-            df = cur_df
-        else:
-            df = pd.concat([df, cur_df])
-
-    gdp_metadata_df = get_gdp_metadata()
+def _chunk_task(df_chunk: pd.DataFrame, chunk_id: str, pos: int, gdp_metadata_df: pd.DataFrame, config: ParsingConfiguration):
[…]
