### 2024-05-10 #6705 @selipot
> Hi @AnjaliSandip, I'll be happy to review the paper :) I have never reviewed any JOSS paper, so I'll be happy to get some guidelines (I read the review criteria and checklist webpages, but this does not explicitly say how to proceed with the review)

YES! Thank you for being available for open science!
