### 2024-05-10 5e4f752
edits to the texts
