### 2024-05-10 5ba4604
🔍 update dev install instructions (#438)
* update dev install instructions

* edits to the texts

---------

Co-authored-by: Shane Elipot <selipot@miami.edu>
