+++ README.md
@@ -111,17 +111,45 @@ If you need the latest development version, you can install it directly from thi
 
 ##### pip:
 
-In your virtual environment, type:
+In your existing virtual environment, you can use `pip` as follows.
+1. Get the code:
 
+```bash
+git clone https://github.com/cloud-drift/clouddrift
+cd clouddrift/
 ```
-pip install git+https://github.com/cloud-drift/clouddrift
+
+2. Install dependencies and local version of `clouddrift`:
+```bash
+pip install .
 ```
 
 ##### Conda:
+
+Using `conda`, you can proceed as follows.
+
+1. Get the code:
+
+```bash
+git clone https://github.com/cloud-drift/clouddrift
+cd clouddrift/
 ```
-conda env create -f environment.yml
+
+2. Create an environment as specified in the yml file with the required library dependencies:
+```bash
+conda env create -f environment.yml # creates a new env with the dependencies
+conda env update -f environment.yml # install dependencies in current environment
 ```
-with the environment [file](https://github.com/Cloud-Drift/clouddrift/blob/main/environment.yml) located in the main repository.
+
+2a. Make sure you **created** the environment by activating it:
+```bash
+conda activate clouddrift
+```
+
+3. Finally, install the local version of `clouddrift`:
+```bash
+pip install .
+```
 
 ### Installing `clouddrift` on unsupported platforms
 
