### 2024-05-10 be09f4d
🔧 configure tests paths in pyproject.toml
