### 2024-05-10 83945ac
🔧 configure tests paths in pyproject.toml (#441)
