### 2024-05-10 742cce9
generalize test discovery and fix test path issue
