+++ tests/adapters/gdp/gdp6h_tests.py
@@ -2,7 +2,7 @@ import os
 import unittest
 from unittest.mock import Mock, patch
 
-import clouddrift.adapters.gdp6h as gdp6h
+import clouddrift.adapters as gdp6h
 from tests.adapters.utils import MultiPatcher
 
 
