+++ pyproject.toml
@@ -75,7 +75,9 @@ files = [
 ]
 
 [tool.pytest.ini_options]
-testpaths = ["tests/*_tests.py", "tests/adapters/*_tests.py"]
+testpaths = [
+  "tests/**/*_tests.py",
+]
 
 
 [[tool.mypy.overrides]]
