### 2024-05-10 #6742 @editorialbot
[REVIEW]: Clouddrift: a Python package to accelerate the use of Lagrangian data for atmospheric, oceanic, and climate sciences
**Submitting author:** @selipot (<a href="http://orcid.org/0000-0001-6051-5426">Shane Elipot</a>)
**Repository:** https://github.com/Cloud-Drift/clouddrift
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.38.1
