### 2024-05-10 be4ee71
chore: Convert xarray variables to numpy arrays for consistent operations in kinematics.velocity_from_position and kinematics.position_from_velocity
