### 2024-05-10 e865218
concat chunked traj rows and merge as aggregate ds
