### 2024-05-14 1541444
rework parallelization
