+++ clouddrift/adapters/gdp/rawfiles.py
@@ -3,18 +3,19 @@ from __future__ import annotations
 import asyncio
 import datetime
 import logging
-import multiprocessing as mp
 import os
+import shutil
 import tempfile
+import warnings
 from collections import defaultdict
+from concurrent.futures import Future, ProcessPoolExecutor, as_completed, wait
 from dataclasses import dataclass
-from multiprocessing.pool import AsyncResult
 from typing import Callable, Literal
 
 import numpy as np
 import pandas as pd
 import xarray as xr
-from tqdm import tqdm
+from tqdm.asyncio import tqdm
 
 from clouddrift.adapters.gdp import get_gdp_metadata
 from clouddrift.adapters.utils import download_with_progress
@@ -361,10 +362,25 @@ def preprocess(id_, **kwargs) -> xr.Dataset:
     dataset = xr.Dataset(variables, coords=coords)
     return dataset
 
-def _chunk_task(df_chunk: pd.DataFrame, chunk_id: str, gdp_metadata_df: pd.DataFrame, config: ParsingConfiguration):
-    ids = np.unique(df_chunk[["id"]].values)
+
+def _process_chunk(df_chunk: pd.DataFrame, chunk_id: str, gdp_metadata_df: pd.DataFrame, config: ParsingConfiguration):
+    zarr_path = f"{os.path.join(_TMP_PATH, "chunks", f"{chunk_id}")}.zarr"
+
+    # remove the current zar archive if it exists
+    if os.path.exists(zarr_path):
+        shutil.rmtree(zarr_path)
+    
+    ids_in_data = np.unique(df_chunk[['id']].values)
+    ids_with_md = np.intersect1d(ids_in_data, gdp_metadata_df[["ID"]].values)
+
+    if len(ids_in_data) > len(ids_with_md):
+        warnings.warn(
+            "Data has drifter ids not found in the metadata table. "
+            + "Please inspect the following ids"
+        )
+
     ra = RaggedArray.from_files(
-        indices=ids,
+        indices=ids_with_md,
         preprocess_func=preprocess,
         rowsize_func=rowsize,
         name_coords=config.coords,
@@ -376,9 +392,106 @@ def _chunk_task(df_chunk: pd.DataFrame, chunk_id: str, gdp_metadata_df: pd.DataF
         config=config,
         tqdm=dict(disable=True)
     )
-    zarr_path = f"{os.path.join(_TMP_PATH, "chunks", chunk_id)}.zarr"
     ra.to_xarray().to_zarr(zarr_path)
-    return zarr_path, ids
+    return zarr_path, ids_with_md
+
+
+def _combine_chunked_drifter_datasets(datasets: list[xr.Dataset]):
+    new_rowsize = sum([ds.rowsize.values[0] for ds in datasets])
+    traj_dataset = xr.concat(datasets, dim="obs", coords="minimal", data_vars=_DATA_VARS, compat="override")
+    traj_dataset["rowsize"] = xr.DataArray(np.array([new_rowsize], dtype=np.int64), coords=traj_dataset["rowsize"].coords)
+    return traj_dataset
+
+
+async def _parallel_get(
+    sources: list[str],
+    gdp_metadata_df: pd.DataFrame,
+    config: ParsingConfiguration,
+    chunk_size: int,
+) -> list[xr.Dataset]:
+
+    with ProcessPoolExecutor(max_workers=os.cpu_count() // 2) as ppe:
+        drifter_chunked_datasets: dict[int, list[xr.Dataset]] = defaultdict(list)
+        for fp in tqdm(
+            sources,
+            desc="Loading files",
+            unit="file",
+            ncols=80,
+            total=len(sources),
+            position=0
+        ):
[…]
