+++ tests/adapters/gdp/gdp6h_tests.py
@@ -2,7 +2,7 @@ import os
 import unittest
 from unittest.mock import Mock, patch
 
-import clouddrift.adapters as gdp6h
+from clouddrift.adapters import gdp6h
 from tests.adapters.utils import MultiPatcher
 
 
