+++ clouddrift/adapters/gdp/rawfiles.py
@@ -27,7 +27,7 @@ _TMP_PATH = os.path.join(tempfile.gettempdir(), "clouddrift", "gdpraw")
 _FILNAME_TEMPLATE = "buoydata_{start}_{end}_{suffix}.dat.gz"
 _SECONDS_IN_DAY = 86_400
 
-_DATA_VARS= [
+_DATA_VARS = [
     "lat",
     "lon",
     "drogue",
@@ -130,11 +130,14 @@ class ParsingConfiguration:
 def _future_years_mask(df):
     return df["obsYear"] > datetime.datetime.now().year
 
+
 def _future_years_mask_sen(df):
     return df["posObsYear"] > datetime.datetime.now().year
 
+
 def _future_years_mask_pos(df):
     return df["senObsYear"] > datetime.datetime.now().year
+
 
 def _bad_drogue_values_mask(df):
     return df["drogue"].astype(np.str_).str.match(r"(\d+[\.]+){2,}")
@@ -160,10 +163,7 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
                     _parse_datetime_with_day_ratio,
                 )
             },
-            remove=[
-                _future_years_mask,
-                _bad_drogue_values_mask
-            ],
+            remove=[_future_years_mask, _bad_drogue_values_mask],
             coords=["id", "obsDatetime"],
         ),
         "sensor": ParsingConfiguration(
@@ -198,10 +198,7 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
                     _parse_datetime_with_day_ratio,
                 )
             },
-            remove=[
-                _future_years_mask,
-                _bad_drogue_values_mask
-            ],
+            remove=[_future_years_mask, _bad_drogue_values_mask],
             coords=["id", "obsDatetime"],
         ),
         "both": ParsingConfiguration(
@@ -256,7 +253,7 @@ def _get_parsing_config(kind: _RecordKind) -> ParsingConfiguration:
             remove=[
                 _future_years_mask_pos,
                 _future_years_mask_sen,
-                _bad_drogue_values_mask
+                _bad_drogue_values_mask,
             ],
             coords=["id", "posObsDatetime", "sensorObsDatetime"],
         ),
@@ -306,10 +303,7 @@ def preprocess(id_, **kwargs) -> xr.Dataset:
     rowsize = len(var)
 
     variables = {
-        "rowsize": (
-            ["traj"],
-            np.array([rowsize], dtype=np.int64)
-        ),
+        "rowsize": (["traj"], np.array([rowsize], dtype=np.int64)),
         "wmo_number": (
             ["traj"],
             traj_md_df[["WMO_number"]].values[0].astype(np.int64),
@@ -363,15 +357,20 @@ def preprocess(id_, **kwargs) -> xr.Dataset:
     return dataset
 
 
-def _process_chunk(df_chunk: pd.DataFrame, chunk_id: str, gdp_metadata_df: pd.DataFrame, config: ParsingConfiguration):
+def _process_chunk(
+    df_chunk: pd.DataFrame,
+    chunk_id: str,
+    gdp_metadata_df: pd.DataFrame,
+    config: ParsingConfiguration,
+):
     filename = f"{chunk_id}.zarr"
     zarr_path = os.path.join(_TMP_PATH, "chunks", filename)
 
     # remove the current zar archive if it exists
     if os.path.exists(zarr_path):
         shutil.rmtree(zarr_path)
-    
-    ids_in_data = np.unique(df_chunk[['id']].values)
+
+    ids_in_data = np.unique(df_chunk[["id"]].values)
     ids_with_md = np.intersect1d(ids_in_data, gdp_metadata_df[["ID"]].values)
 
     if len(ids_in_data) > len(ids_with_md):
[…]
