### 2024-05-14 dfdff12
metadata needs dir for md created
