### 2024-05-14 b876f70
fix tmp_path issues in rawfiles and other datasets
