+++ clouddrift/datasets.py
@@ -159,7 +159,6 @@ def gdp6h(decode_times: bool = True) -> xr.Dataset:
 
 
 def gdpraw(
-    decode_times: bool = True,
     kind: _RecordKind = "raw",
     tmp_path: str = adapters.rawfiles._TMP_PATH,
     max: int | None = None,
@@ -224,7 +223,11 @@ def glad(decode_times: bool = True) -> xr.Dataset:
     return _dataset_filecache("glad.nc", decode_times, adapters.glad.to_xarray)
 
 
-def hurdat2(basin: _BasinOption = "both", decode_times: bool = True) -> xr.DataArray:
+def hurdat2(
+    basin: _BasinOption = "both",
+    decode_times: bool = True,
+    tmp_path: str = adapters.hurdat2._DEFAULT_FILE_PATH,
+) -> xr.DataArray:
     """Returns the revised hurricane database (HURDAT2) as a ragged array xarray dataset.
 
     The function will first look for the ragged array dataset on the local
@@ -298,7 +301,7 @@ def hurdat2(basin: _BasinOption = "both", decode_times: bool = True) -> xr.DataA
     return _dataset_filecache(
         f"hurdat2_{basin}.nc",
         decode_times,
-        lambda: adapters.hurdat2.to_raggedarray(basin).to_xarray(),
+        lambda: adapters.hurdat2.to_raggedarray(basin, tmp_path).to_xarray(),
     )
 
 
