### 2024-05-15 895e98e
sort joined drifter datasets and vars
