### 2024-05-17 d3339d3
sort drifter vars using obsPosDateTime
