### 2024-05-18 1224972
chore: Convert xarray variables to numpy arrays for consistent operations in kinematics.velocity_from_position and kinematics.position_from_velocity (#443)
