### 2024-05-20 44b2b2e
fix sorting and only store aggregate file
