### 2024-05-20 5fb4c8f
index entries using their row index, increased processing performance
