+++ clouddrift/adapters/gdp/rawfiles.py
@@ -27,6 +27,11 @@ _TMP_PATH = os.path.join(tempfile.gettempdir(), "clouddrift", "gdpraw")
 _FILENAME_TEMPLATE = "buoydata_{start}_{end}_{suffix}.dat.gz"
 _SECONDS_IN_DAY = 86_400
 
+_COORDS = [
+    "id",
+    "obs_index"
+]
+
 _DATA_VARS = [
     "lat",
     "lon",
@@ -58,6 +63,33 @@ _RecordKind = Literal["position"] | Literal["sensor"] | Literal["raw"]
 _logger = logging.getLogger(__name__)
 
 
+@dataclass
+class ParsingConfiguration:
+    cols: list[str]
+    col_dtypes: dict[str, type]
+    remove: list[Callable[[pd.DataFrame], pd.DataFrame]]
+    transform: dict[str, tuple[list[str], Callable[..., pd.DataFrame]]]
+
+    def apply_remove(self, df: pd.DataFrame):
+        temp_df = df
+        for filter_ in self.remove:
+            mask = filter_(temp_df)
+            temp_df = temp_df[~mask]
+        return temp_df
+
+    def apply_transform(self, df: pd.DataFrame):
+        tmp_df = df
+        for output_col in self.transform.keys():
+            input_cols, func = self.transform[output_col]
+            args = list()
+            for col in input_cols:
+                arg = df[[col]].values.flatten()
+                args.append(arg)
+            tmp_df = tmp_df.assign(**{output_col: func(*args)})
+            tmp_df = tmp_df.drop(input_cols, axis=1)
+        return tmp_df
+
+
 def _parse_datetime_with_day_ratio(
     month_series: np.ndarray, day_series: np.ndarray, year_series: np.ndarray
 ):
@@ -74,70 +106,12 @@ def _parse_datetime_with_day_ratio(
     return np.array(values, dtype="datetime64[ns]")
 
 
-@dataclass
-class ParsingConfiguration:
-    cols: list[str]
-    coords: list[str]
-    col_dtypes: dict[str, type]
-    remove: list[Callable[[pd.DataFrame], pd.DataFrame]]
-    transform: dict[str, tuple[str, list[str], Callable[..., pd.DataFrame]]]
-    sortCoord: str
-
-    def get_vars_config_map(self, dim_name: str, df: pd.DataFrame):
-        all_config_map = self._get_all_config_map(dim_name, df)
-        return {
-            col: all_config_map[col]
-            for col in all_config_map.keys()
-            if col not in self.coords
-        }
-
-    def get_coords_config_map(self, dim_name: str, df: pd.DataFrame):
-        all_config_map = self._get_all_config_map(dim_name, df)
-        return {col: all_config_map[col] for col in self.coords}
-
-    def _get_all_config_map(self, dim_name: str, df: pd.DataFrame):
-        post_remove = self._apply_remove(df)
-        pre_transform = dict()
-
-        for col in self.cols:
-            data_array = (
-                post_remove[[col]].to_numpy().flatten().astype(self.col_dtypes[col])
-            )
-            pre_transform[col] = ([dim_name], data_array)
-        post_transform = self._apply_transform(pre_transform)
-        return post_transform
-
-    def _apply_remove(self, df: pd.DataFrame):
-        temp_df = df
-        for filter_ in self.remove:
-            mask = filter_(temp_df)
-            temp_df = temp_df[~mask]
-        return temp_df
-
-    def _apply_transform(
-        self, variable_config_map: dict[str, tuple[list[str], pd.DataFrame]]
-    ):
-        for output_col in self.transform.keys():
-            dim, input_cols, func = self.transform[output_col]
-            args = list()
-            for col in input_cols:
-                _, variable = variable_config_map[col]
-                args.append(variable)
[…]
