### 2024-05-21 af734ac
more filtering and improved warnings
