### 2024-05-21 682f806
add datetime vars and correctly type them
