### 2024-05-22 2e2c3d2
Improve documentation, cleanup impl, ruff and mypy
