+++ clouddrift/adapters/gdp/rawfiles.py
@@ -10,7 +10,7 @@ import warnings
 from collections import defaultdict
 from concurrent.futures import Future, ProcessPoolExecutor, as_completed
 from dataclasses import dataclass
-from typing import Callable, Literal
+from typing import Callable
 
 import numpy as np
 import pandas as pd
@@ -27,14 +27,11 @@ _TMP_PATH = os.path.join(tempfile.gettempdir(), "clouddrift", "gdpraw")
 _FILENAME_TEMPLATE = "buoydata_{start}_{end}_{suffix}.dat.gz"
 _SECONDS_IN_DAY = 86_400
 
-_COORDS = [
-    "id",
-    "obs_index"
-]
+_COORDS = ["id", "obs_index"]
 
 _DATA_VARS = [
-    "lat",
-    "lon",
+    "latitude",
+    "longitude",
     "drogue",
     "sst",
     "voltage",
@@ -58,7 +55,94 @@ _METADATA_VARS = [
     "death_code",
 ]
 
-_RecordKind = Literal["position"] | Literal["sensor"] | Literal["raw"]
+var_attrs = {
+    "id": {"long_name": "Global Drifter Program Buoy ID", "units": "-"},
+    "rowsize": {
+        "long_name": "Number of observations per trajectory",
+        "sample_dimension": "obs",
+        "units": "-",
+    },
+    "wmo_number": {
+        "long_name": "World Meteorological Organization buoy identification number",
+        "units": "-",
+    },
+    "current_program": {
+        "long_name": "Current Program",
+        "units": "-",
+        "_FillValue": "-1",
+    },
+    "buoys_type": {
+        "long_name": "Buoy type (see https://www.aoml.noaa.gov/phod/dac/dirall.html)",
+        "units": "-",
+    },
+    "start_date": {
+        "long_name": "First good date and time derived by DAC quality control",
+        "units": "seconds since 1970-01-01 00:00:00",
+    },
+    "start_lon": {
+        "long_name": "First good longitude derived by DAC quality control",
+        "units": "degrees_east",
+    },
+    "start_lat": {
+        "long_name": "Last good latitude derived by DAC quality control",
+        "units": "degrees_north",
+    },
+    "end_date": {
+        "long_name": "Last good date and time derived by DAC quality control",
+        "units": "seconds since 1970-01-01 00:00:00",
+    },
+    "end_lon": {
+        "long_name": "Last good longitude derived by DAC quality control",
+        "units": "degrees_east",
+    },
+    "end_lat": {
+        "long_name": "Last good latitude derived by DAC quality control",
+        "units": "degrees_north",
+    },
+    "drogue_off_date": {
+        "long_name": "Date and time of drogue loss",
+        "units": "seconds since 1970-01-01 00:00:00",
+    },
+    "death_code": {
+        "long_name": "Type of death",
+        "units": "-",
+        "comments": "0 (buoy still alive), 1 (buoy ran aground), 2 (picked up by vessel), 3 (stop transmitting), 4 (sporadic transmissions), 5 (bad batteries), 6 (inactive status)",
+    },
+    "longitude": {"long_name": "Longitude", "units": "degrees_east"},
+    "latitude": {"long_name": "Latitude", "units": "degrees_north"},
+    "drogue": {
+        "long_name": "Status indicating the presence of the drogue",
+        "units": "-",
+        "flag_values": "1,0",
+        "flag_meanings": "drogued, undrogued",
+    },
+    "sst": {
+        "long_name": "Fitted sea water temperature",
+        "units": "Kelvin",
[…]
