### 2024-06-02 #6742 @malmans2
Hi @selipot,

I'll try to review this submission this week, sorry about the delay.
I'll open issues if I have recommendations, and will provide a final recap here when I'm done.
One question about the authors: Rick Lumpkin is the only author that did not contribute to the code on GitHub.
Could you please clarify how they contributed to Clouddrift?

Thanks!
