### 2024-06-02 Cloud-Drift/clouddrift#446 @malmans2
⭐ Improve installation of optional dependencies
In the README there are instructions on how to manually install optional dependencies.
`setuptools` provides a way to manage optional dependencies. For example, you can easily achieve something like this:
```
pip install clouddrift[plotting]
pip install clouddrift[complete]
```

All you need to do is adding a section in you pyproject.toml that looks like this:
```toml
[project.optional-dependencies]
complete = ["clouddrift[plotting,foo]"]
plotting = ["matplotlib", "cartopy"]
foo = ["foo", "bar"]
```

This should be the relevant documentation: https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#

JOSS: https://github.com/openjournals/joss-reviews/issues/6742
