### 2024-06-03 Cloud-Drift/clouddrift#447 @malmans2
🔎 Add section with dependencies
I noticed that the documentation is missing a section listing all dependencies.
Including this information would be very helpful for users. For example, you could add this information in the installation section, similar to how it is done by xarray: https://docs.xarray.dev/en/stable/getting-started-guide/installing.html

JOSS: https://github.com/openjournals/joss-reviews/issues/6742
