### 2024-06-03 Cloud-Drift/clouddrift#450 @malmans2
🐛 Unit test test_subsurface_floats_opens fails
**Describe the bug**
I get a failure when I run the unit test suite

**To Reproduce**
```
git clone https://github.com/cloud-drift/clouddrift
cd clouddrift/
mamba env create -f environment.yml --force
conda activate clouddrift
mamba install -c conda-forge matplotlib cartopy
pip install -e .
python -m unittest tests/*.py
```

 - MacOS
 - git+https://github.com/cloud-drift/clouddrift@12249729bb1f4ce02cd78663c35c88c58766765f

**Additional context**
```
ERROR: test_subsurface_floats_opens (tests.datasets_tests.datasets_tests.test_subsurface_floats_opens)
Traceback (most recent call last):
  File "/Users/mattia/MyGit/JOSS/clouddrift/tests/datasets_tests.py", line 47, in test_subsurface_floats_opens
    with datasets.subsurface_floats() as ds:
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mattia/MyGit/JOSS/clouddrift/clouddrift/datasets.py", line 484, in subsurface_floats
    return _dataset_filecache(
           ^^^^^^^^^^^^^^^^^^^
  File "/Users/mattia/MyGit/JOSS/clouddrift/clouddrift/datasets.py", line 630, in _dataset_filecache
    ds = get_ds()
         ^^^^^^^^
  File "/Users/mattia/MyGit/JOSS/clouddrift/clouddrift/adapters/subsurface_floats.py", line 49, in to_xarray
    source_data = scipy.io.loadmat(local_file)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mattia/miniforge3/envs/clouddrift/lib/python3.12/site-packages/scipy/io/matlab/_mio.py", line 227, in loadmat
    matfile_dict = MR.get_variables(variable_names)
                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mattia/miniforge3/envs/clouddrift/lib/python3.12/site-packages/scipy/io/matlab/_mio5.py", line 330, in get_variables
    res = self.read_var_array(hdr, process)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/mattia/miniforge3/envs/clouddrift/lib/python3.12/site-packages/scipy/io/matlab/_mio5.py", line 290, in read_var_array
    return self._matrix_reader.array_from_header(header, process)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "_mio5_utils.pyx", line 665, in scipy.io.matlab._mio5_utils.VarReader5.array_from_header
  File "_mio5_utils.pyx", line 710, in scipy.io.matlab._mio5_utils.VarReader5.array_from_header
  File "_mio5_utils.pyx", line 885, in scipy.io.matlab._mio5_utils.VarReader5.read_cells
  File "_mio5_utils.pyx", line 663, in scipy.io.matlab._mio5_utils.VarReader5.read_mi_matrix
  File "_mio5_utils.pyx", line 694, in scipy.io.matlab._mio5_utils.VarReader5.array_from_header
  File "_mio5_utils.pyx", line 768, in scipy.io.matlab._mio5_utils.VarReader5.read_real_complex
  File "_mio5_utils.pyx", line 445, in scipy.io.matlab._mio5_utils.VarReader5.read_numeric
  File "_mio5_utils.pyx", line 350, in scipy.io.matlab._mio5_utils.VarReader5.read_element
  File "_streams.pyx", line 171, in scipy.io.matlab._streams.ZlibInputStream.read_string
  File "_streams.pyx", line 164, in scipy.io.matlab._streams.ZlibInputStream.read_into
OSError: could not read bytes

Ran 216 tests in 34.575s

FAILED (errors=1)
```

JOSS: https://github.com/openjournals/joss-reviews/issues/6742
