### 2024-06-03 7a070e2
added optional dep in pyproject.toml
