### 2024-06-03 Cloud-Drift/clouddrift#450 @malmans2
Side note - not a requirement for JOSS:
Do you perform a checksum validation on the test data?
If not, I suggest looking into [pooch](https://www.fatiando.org/pooch/latest/). Many SciPy packages use it, and in my experience, it works well out of the box. Additionally, I'm personally not a big fan of test data being downloaded to a hidden directory in the home. By default,`pooch` uses the standard machine-specific cache directory.
