### 2024-06-03 Cloud-Drift/clouddrift#450 @philippemiron
1. We don't do checksums at the moment. We could, but that would break tests and require more work when the underlying data (that we don't control) is updated.
2. I did not implement the download module, but looks like `pooch` could be used instead.
3. Not a fan of the hidden folder also personally haha. We do want to be able to save it locally and not in a temporary cache folder to avoid having to redownload the ~20k files of the GDP datasets.
