### 2024-06-03 Cloud-Drift/clouddrift#450 @malmans2
👍
Sorry, too much focus on pooch in my comment. My main point was, maybe you can't reproduce because we are using different test data? I'll try again tomorrow.
