### 2024-06-03 Cloud-Drift/clouddrift#450 @philippemiron
Ah ok. The tests should always pickup the complete datasets from their respective repos. There is sometime issues with timeout but right now it seems to be ok. Let me know when you give it another go.
