### 2024-06-04 Cloud-Drift/clouddrift#450 @malmans2
It worked this morning, same environment. I guess it's either a flake or the data downloaded was corrupted.

> We don't do checksums at the moment. We could, but that would break tests and require more work when the underlying data (that we don't control) is updated.

Maybe the best solution is to store a copy of the data that you need for unit testing where you have full control, like a dedicated GitHub repo, Zenodo, a cloud provider, ...
