### 2024-06-04 939624b
Update install.rst
Co-authored-by: Mattia Almansi <m.almansi@bopen.eu>
