### 2024-06-05 #6742 @malmans2
Hi everyone,

I've been reviewing this submission for the past few days and have opened issues as needed. Here is a summary of the outstanding issues:

1. **Contribution and Authorship**: Needs clarification. See [this comment](https://github.com/openjournals/joss-reviews/issues/6742#issuecomment-2143772957).
2. **Functionality/Performance/Example Usage/Functionality Documentation**: My goal was to assess these points by running the examples and demo notebooks. However, the example usage and the notebook I tried did not work for me. See issues [#449](https://github.com/Cloud-Drift/clouddrift/issues/449) and [#452](https://github.com/Cloud-Drift/clouddrift/issues/452).
3. **Installation Instructions**: The dependencies upon which Clouddrift is built are not clearly mentioned anywhere. See issue [#446](https://github.com/Cloud-Drift/clouddrift/issues/446).
4. **State of the Field**: I don't see a discussion about the state of the field in the paper. Is Clouddrift unique and cannot be compared to other libraries, whether Python or otherwise? This is a genuine question, as I'm not an expert in Lagrangian data at all.
5. **References**: Xarray is the only library mentioned in the paper, but I see a dozen of mandatory dependencies. Is Clouddrift meant to be an extension of Xarray? If yes, I suggest making it clearer in the paper. If no, I think more core dependencies need to be mentioned and cited.

I will finish my review when the example code/notebooks are fixed. If I was unlucky and picked the wrong notebook, please let me know which notebooks I should try.
