### 2024-06-06 2c20247
added dev packages and complete to install everything
