+++ pyproject.toml
@@ -4,7 +4,11 @@ build-backend = "hatchling.build"
 
 [project]
 name = "clouddrift"
+<<<<<<< HEAD
 version = "0.36.0"
+=======
+version = "0.35.1"
+>>>>>>> eedeed5 (add default url and path to download function (#454))
 authors = [
   { name="Shane Elipot", email="selipot@miami.edu" },
   { name="Philippe Miron", email="philippemiron@gmail.com" },
