### 2024-06-06 eedeed5
add default url and path to download function (#454)
Co-authored-by: Philippe Miron <philippe.miron@dtn.com>
