### 2024-06-06 190e800
add optional deps to documentation
