### 2024-06-06 52f02df
small modifications to docs
