### 2024-06-07 Cloud-Drift/clouddrift#456 @rcaneill
🐛 binder mosaic is broken
**Describe the bug**

Input cell [5] raises an error:

```
from clouddrift.analysis import apply_ragged, velocity_from_position
```

```
Cell In[5], line 1
----> 1 from clouddrift.analysis import apply_ragged, velocity_from_position

ModuleNotFoundError: No module named 'clouddrift.analysis'
```

I guess that some modules have changed name recently
