### 2024-06-07 #6742 @rcaneill
Hi,

I did my review, here are the comments:

* **Contribution and authorship**: @selipot made major contributions. However, I do not see any commit from Rick Lumpkin
* **Reproducibility**: importing numpy was missing in the paper
  https://github.com/Cloud-Drift/clouddrift/pull/455
* **Example usage**: apart from the comment that @malmans2 [made](https://github.com/Cloud-Drift/clouddrift/issues/449), the mosaic example is also broken: https://github.com/Cloud-Drift/clouddrift/issues/456
  I believe that the examples notebooks are very good and full of information and should also be included into the main documentation
* **References**: I think that xarray should be properly cited, and not only mentioned with an url
* **Installation instruction**: It seems that the deps are handled by hand. The devs deps are found in file CONTRIBUTING.md (will be solved by https://github.com/Cloud-Drift/clouddrift/pull/451)

As @malmans2 , I'll have a proper look when the binder notebooks will be fixed. I did not re-write some comments I had that @malmans2 already covered, and I moved the comments I initially made on the checklist comment in this comment (I had misunderstood how to proceed)
