### 2024-06-07 eabdf10
numpy is missing in the example (#455)
