### 2024-06-10 Cloud-Drift/clouddrift#452 @philippemiron
Once we release a new version, it should be easier to fix with https://github.com/Cloud-Drift/clouddrift/pull/454.
