### 2024-06-11 19c16ef
⭐ optional dep in pyproject.toml (#451)
