+++ docs/requirements.txt
@@ -1,3 +1,0 @@
-sphinx
-sphinx-book-theme
-sphinx-copybutton
