### 2024-06-11 1a97d7a
🔎 install plotting dep update in README.md
