### 2024-06-12 Cloud-Drift/clouddrift#449 @malmans2
OK, the usage example looks good now.
Consider running doctests in the CI. I tried it, and it needs some fix here and there:
`pytest --doctest-glob="*.rst" docs/`
