### 2024-06-13 #6742 @selipot
> Hi @selipot,
>
> I'll try to review this submission next week, sorry about the delay. I'll open issues if I have recommendations, and will provide a final recap here once I'm done. I have one question regarding the authorship: Rick Lumpkin is the only author who did not contribute to the code on GitHub. Could you please clarify how they contributed to Clouddrift?
>
> Thanks!

While Rick Lumpkin did not commit directly to the repo, he is an official collaborator of the NSF grant which is funding this work. He has consistently joined our weekly meeting to discuss strategies, datasets, and functionality of the software. As such, he is an integral part of the team that produced this software.
