+++ paper.bib
@@ -35,3 +35,58 @@ author = {Erik {van Sebille} and Stephen M. Griffies and Ryan Abernathey and Tho
   url = {https://doi.org/10.25921/x46c-3620},
   doi = {10.25921/x46c-3620}
 }
+
+@article{Delandmeter:2019,
+AUTHOR = {Delandmeter, P. and van Sebille, E.},
+TITLE = {The Parcels v2.0 Lagrangian framework: new field interpolation schemes},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {12},
+YEAR = {2019},
+NUMBER = {8},
+PAGES = {3571--3584},
+URL = {https://gmd.copernicus.org/articles/12/3571/2019/},
+DOI = {10.5194/gmd-12-3571-2019}
+}
+
+@article{Stein:2015,
+  title={NOAA’s HYSPLIT atmospheric transport and dispersion modeling system},
+  author={Stein, Ariel F and Draxler, Roland R and Rolph, Glenn D and Stunder, Barbara JB and Cohen, Mark D and Ngan, Fong},
+  journal={Bulletin of the American Meteorological Society},
+  volume={96},
+  number={12},
+  pages={2059--2077},
+  year={2015},
+  publisher={American Meteorological Society},
+  DOI = {https://doi.org/10.1175/BAMS-D-14-00110.1}
+}
+
+@article{Maze:2020, 
+  doi = {10.21105/joss.02425}, 
+  url = {https://doi.org/10.21105/joss.02425}, 
+  year = {2020}, 
+  publisher = {The Open Journal}, 
+  volume = {5}, 
+  number = {53}, 
+  pages = {2425}, 
+  author = {Guillaume Maze and Kevin Balem}, 
+  title = {argopy: A Python library for Argo ocean data analysis}, 
+  journal = {Journal of Open Source Software} 
+}
+
+@article{Lilly:2021,
+  doi = {10.5281/zenodo.4547006},
+  url = {http://www.jmlilly.net/software},
+  year = {2021},
+  author = {Jonathan M. Lilly},
+  title = {jLab: A data analysis package for Matlab},
+}
+
+@article{Hoyer:2017,
+  title={xarray: ND labeled arrays and datasets in Python},
+  author={Hoyer, Stephan and Hamman, Joe},
+  journal={Journal of Open Research Software},
+  volume={5},
+  number={1},
+  pages={10--10},
+  year={2017}
+}
