+++ paper.md
@@ -48,7 +48,9 @@ Lagrangian data in Earth sciences are unique because they do not conform to esta
 
 # Statement of need
 
-In Earth, Ocean, Geo-, and Atmospheric Science, *Eulerian data* typically refers to a type of data acquired or simulated at a particular fixed point or region in space. Eulerian data are defined on fixed spatiotemporal grids with monotonic coordinates (e.g. latitude, longitude, depth, time) for which popular Python tools such as [Xarray](https://docs.xarray.dev/en/stable/) are naturally suited. In contrast, *Lagrangian data* are acquired by observing platforms that move with the flow they are embedded in, for example uncrewed platforms, vehicles, virtual particles, atmospheric phenomena such as tropical cyclones, and even animals that gather data along their natural but complex paths. Because such paths traverse both spatial and temporal dimensions, Lagrangian data often convolve spatial and temporal information that cannot consistently and readily be organized, cataloged, and stored in common data structures and file formats with the help of common libraries and standards. As an example, the concepts of dimensions and coordinates for Lagrangian data are ambiguous and not clearly established. As such, for both data generators and data users, Lagrangian data present challenges that the Clouddrift Python library aims to overcome. 
+In Earth, Ocean, Geo-, and Atmospheric Science, *Eulerian data* typically refers to a type of data acquired or simulated at a particular fixed point or region in space. Eulerian data are defined on fixed spatiotemporal grids with monotonic coordinates (e.g. latitude, longitude, depth, time) for which popular Python tools such as [Xarray](https://docs.xarray.dev/en/stable/)[@Hoyer:2017] are naturally suited. In contrast, *Lagrangian data* are acquired by observing platforms that move with the flow they are embedded in, for example uncrewed platforms, vehicles, virtual particles, atmospheric phenomena such as tropical cyclones, and even animals that gather data along their natural but complex paths. Because such paths traverse both spatial and temporal dimensions, Lagrangian data often convolve spatial and temporal information that cannot consistently and readily be organized, cataloged, and stored in common data structures and file formats with the help of common libraries and standards. As an example, the concepts of dimensions and coordinates for Lagrangian data are ambiguous and not clearly established. As such, for both data generators and data users, Lagrangian data present challenges that the Clouddrift Python library aims to overcome. 
+
+The Clouddrift library is distinct from other tools which aim at simulating the trajectories of particles in oceanic and atmospheric models, such as [OceanParcels](https://oceanparcels.org) [@Delandmeter:2019], or [HYSPLIT](https://www.ready.noaa.gov/HYSPLIT.php)[@Stein:2015], because this software is focusing on the analysis of the data emanating from observational and numerical Lagrangian experiments, and the curation of Lagrangian datasets, using consistent data structures and approaches (See Scope and key features). Clouddrift, as a library, shares some intentions with argopy [@Maze:2020], a python software that aims at accessing and manipulating the Argo dataset, a specific Lagrangian oceanographic dataset, but clouddrift intends on being dataset-agnostic and ultimately on going beyond solely Earth data. Clouddrift adapts into Python some oceanographic analysis functions from jLab, a data analysis package for Matlab [@Lilly:2021], in agreement with its license.
 
 # Scope and key features
 
