### 2024-06-14 ce71a6a
bibtex formatting
