+++ paper.md
@@ -50,7 +50,7 @@ Lagrangian data in Earth sciences are unique because they do not conform to esta
 
 In Earth, Ocean, Geo-, and Atmospheric Science, *Eulerian data* typically refers to a type of data acquired or simulated at a particular fixed point or region in space. Eulerian data are defined on fixed spatiotemporal grids with monotonic coordinates (e.g. latitude, longitude, depth, time) for which popular Python tools such as [Xarray](https://docs.xarray.dev/en/stable/) [@Hoyer:2017] are naturally suited. In contrast, *Lagrangian data* are acquired by observing platforms that move with the flow they are embedded in, for example uncrewed platforms, vehicles, virtual particles, atmospheric phenomena such as tropical cyclones, and even animals that gather data along their natural but complex paths. Because such paths traverse both spatial and temporal dimensions, Lagrangian data often convolve spatial and temporal information that cannot consistently and readily be organized, cataloged, and stored in common data structures and file formats with the help of common libraries and standards. As an example, the concepts of dimensions and coordinates for Lagrangian data are ambiguous and not clearly established. As such, for both data generators and data users, Lagrangian data present challenges that the clouddrift Python library aims to overcome. 
 
-The clouddrift library is distinct from other tools designed to simulate particle trajectories in oceanic and atmospheric models, such as [OceanParcels](https://oceanparcels.org) [@Delandmeter:2019], or [HYSPLIT](https://www.ready.noaa.gov/HYSPLIT.php)[@Stein:2015]. Unlike these tools, clouddrift focuses on analyzing data from observational and numerical Lagrangian experiments and curating Lagrangian datasets using consistent data structures and methodologies (See Scope and key features). While clouddrift shares some goals with argopy [@Maze:2020], a Python library for accessing and manipulating the Argo dataset (a specific Lagrangian oceanographic dataset), clouddrift aims to be dataset-agnostic and extend beyond just Earth data. Additionally, clouddrift incorporates oceanographic analysis functions from jLab, a Matlab data analysis package [@Lilly:2021], in compliance with its license. Clouddrift core Python dependencies include NumPy [@Harris:2020] and Scipy [@Virtanen:2020] for data analysis, as well as Xarray [@Hoyer:2017], pandas [@reback2020pandas;@mckinney-proc-scipy-2010], and [Awkward Array](https://awkward-array.org/doc/main/index.html) for its data processing and manipulation functions. 
+The clouddrift library is distinct from other tools designed to simulate particle trajectories in oceanic and atmospheric models, such as [OceanParcels](https://oceanparcels.org) [@Delandmeter:2019], or [HYSPLIT](https://www.ready.noaa.gov/HYSPLIT.php) [@Stein:2015]. Unlike these tools, clouddrift focuses on analyzing data from observational and numerical Lagrangian experiments and curating Lagrangian datasets using consistent data structures and methodologies (See Scope and key features). While clouddrift shares some goals with argopy [@Maze:2020], a Python library for accessing and manipulating the Argo dataset (a specific Lagrangian oceanographic dataset), clouddrift aims to be dataset-agnostic and extend beyond just Earth data. Additionally, clouddrift incorporates oceanographic analysis functions from jLab, a Matlab data analysis package [@Lilly:2021], in compliance with its license. Clouddrift core Python dependencies include NumPy [@Harris:2020] and Scipy [@Virtanen:2020] for data analysis, as well as Xarray [@Hoyer:2017], pandas [@reback2020pandas;@mckinney-proc-scipy-2010], and [Awkward Array](https://awkward-array.org/doc/main/index.html) for its data processing and manipulation functions. 
 
 # Scope and key features
 
