### 2024-06-14 8642968
🔍 revision paper.md (#461)
* Update paper.md

* Sort ascending affiliations

---------

Co-authored-by: milancurcic <caomaco@gmail.com>
Co-authored-by: Shane Elipot <selipot@miami.edu>
