+++ paper.md
@@ -18,7 +18,7 @@ authors:
   - name: Milan Curcic
     orcid: 0000-0002-8822-7749
     equal-contrib: true
-    affiliation: 1, 4
+    affiliation: 1, 3
   - name: Kevin Santana
     orcid: 0009-0003-8383-1212
     equal-contrib: true
@@ -26,31 +26,31 @@ authors:
   - name: Rick Lumpkin
     orcid: 0000-0002-6690-1704
     equal-contrib: true
-    affiliation: 3
+    affiliation: 4
 
 affiliations:
- - name: Rosenstiel School of Marine, Atmospheric, and Earth science, University of Miami
+ - name: Rosenstiel School of Marine, Atmospheric, and Earth Science, University of Miami
    index: 1
  - name: Florida State University
    index: 2
- - name: NOAA Atlantic Oceanographic and Meteorological Laboratory
-   index: 3
  - name: Frost Institute for Data Science and Computing, University of Miami
+   index: 3
+ - name: NOAA Atlantic Oceanographic and Meteorological Laboratory
    index: 4
 date: 10 April 2024
 bibliography: paper.bib
 
 ---
-
+ 
 # Summary
 
 Lagrangian data in Earth sciences are unique because they do not conform to established standards related to dimensions, coordinates, and organizational structures. In addition, because they convolve spatial and temporal information, Lagrangian data need specific processing and analysis tools for their scientific and operational use. The clouddrift Python library addresses these challenges by offering tools to process and analyze Lagrangian data with an emphasis on the ragged array representation.
 
 # Statement of need
 
-In Earth, Ocean, Geo-, and Atmospheric Science, *Eulerian data* typically refers to a type of data acquired or simulated at a particular fixed point or region in space. Eulerian data are defined on fixed spatiotemporal grids with monotonic coordinates (e.g. latitude, longitude, depth, time) for which popular Python tools such as [Xarray](https://docs.xarray.dev/en/stable/) [@Hoyer:2017] are naturally suited. In contrast, *Lagrangian data* are acquired by observing platforms that move with the flow they are embedded in, for example uncrewed platforms, vehicles, virtual particles, atmospheric phenomena such as tropical cyclones, and even animals that gather data along their natural but complex paths. Because such paths traverse both spatial and temporal dimensions, Lagrangian data often convolve spatial and temporal information that cannot consistently and readily be organized, cataloged, and stored in common data structures and file formats with the help of common libraries and standards. As an example, the concepts of dimensions and coordinates for Lagrangian data are ambiguous and not clearly established. As such, for both data generators and data users, Lagrangian data present challenges that the clouddrift Python library aims to overcome. 
[…]
