### 2024-06-14 db8de99
bibtex fixes
