+++ paper.bib
@@ -26,7 +26,7 @@ author = {Erik {van Sebille} and Stephen M. Griffies and Ryan Abernathey and Tho
 @Misc{Elipot:2022,
   OPTkey = 	 {},
   author = 	 {Elipot, Shane and Sykulski, Adam and Lumpkin, Rick and Centurioni, Luca and Pazos, Mayra},
-  Ttitle = 	 {Hourly location, current velocity, and temperature collected from {Global Drifter Program} drifters world-wide},
+  title = 	 {Hourly location, current velocity, and temperature collected from {Global Drifter Program} drifters world-wide},
   howpublished = {NOAA National Centers for Environmental Information.},
   OPTmonth = 	 {},
   year = 	 {2022},
@@ -82,13 +82,14 @@ DOI = {10.5194/gmd-12-3571-2019}
 }
 
 @article{Hoyer:2017,
-  title={xarray: ND labeled arrays and datasets in {Python}},
+  title={xarray: {ND} labeled {Arrays} and {Datasets} in {Python}},
   author={Hoyer, Stephan and Hamman, Joe},
   journal={Journal of Open Research Software},
   volume={5},
   number={1},
   pages={10--10},
-  year={2017}
+  year={2017},
+  doi = {https://doi.org/10.5334/jors.148}
 }
 
 @Article{Harris:2020,
@@ -139,10 +140,10 @@ DOI = {10.5194/gmd-12-3571-2019}
 }
 
 @software{reback2020pandas,
-    author       = {The pandas development team},
-    title        = {pandas-dev/pandas: Pandas},
+    author       = {{The pandas development team}},
+    title        = {pandas-dev/pandas: {Pandas}},
     month        = feb,
-    year         = 2020,
+    year         = 2024,
     publisher    = {Zenodo},
     version      = {latest},
     doi          = {10.5281/zenodo.3509134},
