### 2024-06-14 2812dca
fix docs using pytest (#460)
Co-authored-by: Philippe Miron <philippe.miron@dtn.com>
