+++ docs/datasets.rst
@@ -8,42 +8,41 @@ datasets.
 
 >>> from clouddrift.datasets import gdp1h
 >>> ds = gdp1h()
-    <xarray.Dataset>
-    Dimensions:                (traj: 17324, obs: 165754333)
-    Coordinates:
-        ids                    (obs) int64 ...
-        lat                    (obs) float32 ...
-        lon                    (obs) float32 ...
-        time                   (obs) datetime64[ns] ...
-    Dimensions without coordinates: traj, obs
-    Data variables: (12/55)
-        BuoyTypeManufacturer   (traj) |S20 ...
-        BuoyTypeSensorArray    (traj) |S20 ...
-        CurrentProgram         (traj) float64 ...
-        DeployingCountry       (traj) |S20 ...
-        DeployingShip          (traj) |S20 ...
-        DeploymentComments     (traj) |S20 ...
-        ...                     ...
-        sst1                   (obs) float64 ...
-        sst2                   (obs) float64 ...
-        typebuoy               (traj) |S10 ...
-        typedeath              (traj) int8 ...
-        ve                     (obs) float32 ...
-        vn                     (obs) float32 ...
-    Attributes: (12/16)
-        Conventions:       CF-1.6
-        acknowledgement:   Elipot, Shane; Sykulski, Adam; Lumpkin, Rick; Centurio...
-        contributor_name:  NOAA Global Drifter Program
-        contributor_role:  Data Acquisition Center
-        date_created:      2022-12-09T06:02:29.684949
-        doi:               10.25921/x46c-3620
-        ...                ...
-        processing_level:  Level 2 QC by GDP drifter DAC
-        publisher_email:   aoml.dftr@noaa.gov
-        publisher_name:    GDP Drifter DAC
-        publisher_url:     https://www.aoml.noaa.gov/phod/gdp
-        summary:           Global Drifter Program hourly data
-        title:             Global Drifter Program hourly drifting buoy collection
+>>> ds
+<xarray.Dataset> Size: 16GB
+Dimensions:                (traj: 19396, obs: 197214787)
+Coordinates:
+    id                     (traj) int64 155kB ...
+    time                   (obs) datetime64[ns] 2GB ...
+Dimensions without coordinates: traj, obs
+Data variables: (12/59)
+    BuoyTypeManufacturer   (traj) |S20 388kB ...
+    BuoyTypeSensorArray    (traj) |S20 388kB ...
+    CurrentProgram         (traj) float32 78kB ...
+    DeployingCountry       (traj) |S20 388kB ...
+    DeployingShip          (traj) |S20 388kB ...
+    DeploymentComments     (traj) |S20 388kB ...
+    ...                     ...
+    start_lat              (traj) float32 78kB ...
+    start_lon              (traj) float32 78kB ...
+    typebuoy               (traj) |S10 194kB ...
+    typedeath              (traj) int8 19kB ...
+    ve                     (obs) float32 789MB ...
+    vn                     (obs) float32 789MB ...
+Attributes: (12/16)
+    Conventions:       CF-1.6
+    acknowledgement:   Elipot, Shane; Sykulski, Adam; Lumpkin, Rick; Centurio...
+    contributor_name:  NOAA Global Drifter Program
+    contributor_role:  Data Acquisition Center
+    date_created:      2023-09-08T17:05:12.130123
+    doi:               10.25921/x46c-3620
+    ...                ...
+    processing_level:  Level 2 QC by GDP drifter DAC
+    publisher_email:   aoml.dftr@noaa.gov
+    publisher_name:    GDP Drifter DAC
+    publisher_url:     https://www.aoml.noaa.gov/phod/gdp
+    summary:           Global Drifter Program hourly data
+    title:             Global Drifter Program hourly drifting buoy collection
 
 Currently available datasets are:
 
