### 2024-06-19 9588e49
raw to source and docs addition
