### 2024-06-19 be39504
docstring edits
