+++ paper.bib
@@ -159,3 +159,15 @@ DOI = {10.5194/gmd-12-3571-2019}
   editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
   doi       = { 10.25080/Majora-92bf1922-00a }
 }
+
+@article{Stern:2022,
+AUTHOR={Stern, Charles and Abernathey, Ryan and Hamman, Joseph and Wegener, Rachel and Lepore, Chiara and Harkins, Sean and Merose, Alexander},   
+TITLE={{Pangeo Forge: Crowdsourcing Analysis-Ready, Cloud Optimized Data Production}},      
+JOURNAL={Frontiers in Climate},      
+VOLUME={3},           
+YEAR={2022},      
+URL={https://www.frontiersin.org/articles/10.3389/fclim.2021.782909},       
+DOI={10.3389/fclim.2021.782909},      
+ISSN={2624-9553},   
+ABSTRACT={Pangeo Forge is a new community-driven platform that accelerates science by providing high-level recipe frameworks alongside cloud compute infrastructure for extracting data from provider archives, transforming it into analysis-ready, cloud-optimized (ARCO) data stores, and providing a human- and machine-readable catalog for browsing and loading. In abstracting the scientific domain logic of data recipes from cloud infrastructure concerns, Pangeo Forge aims to open a door for a broader community of scientists to participate in ARCO data production. A wholly open-source platform composed of multiple modular components, Pangeo Forge presents a foundation for the practice of reproducible, cloud-native, big-data ocean, weather, and climate science without relying on proprietary or cloud-vendor-specific tooling.}
+}
