### 2024-06-21 #6742 @selipot
> Hi everyone,
>
> I've been reviewing this submission for the past few days and have opened issues as needed. Here is a summary of the outstanding issues:
>
> 1. **Contribution and Authorship**: Needs clarification. See [this comment](https://github.com/openjournals/joss-reviews/issues/6742#issuecomment-2143772957).

This has now been [clarified](https://github.com/openjournals/joss-reviews/issues/6742#issuecomment-2167388713).

> 2. **Functionality/Performance/Example Usage/Functionality Documentation**: My goal was to assess these points by running the examples and demo notebooks. However, the example usage and the notebook I tried did not work for me. See issues [#449](https://github.com/Cloud-Drift/clouddrift/issues/449) and [#452](https://github.com/Cloud-Drift/clouddrift/issues/452).

We reviewed and assessed that the clouddrift-examples notebooks were not the best to showcase the library, having been designed a long time ago. We have therefore archived that repository. In the main README.md of the library we know point to [gdp-get-started](https://github.com/Cloud-Drift/gdp-get-started), [mosaic-get-started](https://github.com/Cloud-Drift/mosaic-get-started), [hurdat2-get-started](https://github.com/Cloud-Drift/hurdat2-get-started), 3 notebooks that are running well, including in a binder environment. In addition, we link to a [more advanced repository](https://registry.opendata.aws/hycom-global-drifters/index.html) with further examples using a large cloud-based Lagrangian dataset.

> 3. **Installation Instructions**: The dependencies upon which Clouddrift is built are not clearly mentioned anywhere. See issue [#446](https://github.com/Cloud-Drift/clouddrift/issues/446).

This issue has now been resolved.

> 4. **State of the Field**: I don't see a discussion about the state of the field in the paper. Is Clouddrift unique and cannot be compared to other libraries, whether Python or otherwise? This is a genuine question, as I'm not an expert in Lagrangian data at all.

To address this concern we have now added a new paragraph in the text of the paper (lines 31-45).

> 5. **References**: Xarray is the only library mentioned in the paper, but I see a dozen of mandatory dependencies. Is Clouddrift meant to be an extension of Xarray? If yes, I suggest making it clearer in the paper. If no, I think more core dependencies need to be mentioned and cited.

This has now been addressed in lines 31-45 of the revised paper.

> I will finish my review when the example code/notebooks are fixed. If I was unlucky and picked the wrong notebook, please let me know which notebooks I should try.

The three notebooks mentioned above are now fixed.
