### 2024-06-21 #6742 @selipot
> I did my review, here are the comments:
>
> * **Contribution and authorship**: @selipot made major contributions. However, I do not see any commit from Rick Lumpkin

The contributions of Rick Lumpkin to the paper and software have now been clarified.

> * **Reproducibility**: importing numpy was missing in the paper
>   [🐛 fix missing numpy in the paper's example Cloud-Drift/clouddrift#455](https://github.com/Cloud-Drift/clouddrift/pull/455)

Now fixed.

> * **Example usage**: apart from the comment that @malmans2 [made](https://github.com/Cloud-Drift/clouddrift/issues/449), the mosaic example is also broken: [🐛 binder mosaic is broken Cloud-Drift/clouddrift#456](https://github.com/Cloud-Drift/clouddrift/issues/456)
>   I believe that the examples notebooks are very good and full of information and should also be included into the main documentation

These notebooks are now fixed.  In a future version of the documentation we will probably include the notebooks but prefer to dissociate the ones mentioned above from the library at the moment.

> * **References**: I think that xarray should be properly cited, and not only mentioned with an url

Xarray is now properly cited, along with other core dependencies.

> * **Installation instruction**: It seems that the deps are handled by hand. The devs deps are found in file CONTRIBUTING.md (will be solved by [⭐ optional dep in pyproject.toml Cloud-Drift/clouddrift#451](https://github.com/Cloud-Drift/clouddrift/pull/451))

The installation instructions have now been updated.

> As @malmans2 , I'll have a proper look when the binder notebooks will be fixed. I did not re-write some comments I had that @malmans2 already covered, and I moved the comments I initially made on the checklist comment in this comment (I had misunderstood how to proceed)
