### 2024-06-24 2270510
Use different var names for pos/sensor datetimes
