### 2024-06-24 ae71ebe
🔍 reference latest DOI (#479)
Instead of referencing the DOI related to a release of clouddrift we should use the latest DOI which references the latest release of clouddrift, whatever version number that may be.
