### 2024-06-24 #6742 @selipot
> Thanks @selipot for addressing my comments.
>
> I think something is still not quite right with the authorship and Zenodo:
>
> 1. The paper authors are: Elipot, Miron, Curcic, Santana, and Lumpkin.
> 2. Lumpkin is missing in the [.zenodo.json](https://github.com/Cloud-Drift/clouddrift/blob/main/.zenodo.json).
> 3. Lumpkin is missing in the [version linked in the README](https://zenodo.org/records/11081648), and there is an additional author: Vadim Bertrand.
> 4. Lumpkin is missing in the [latest version on Zenodo](https://zenodo.org/records/11584111), and the version in the title is not correct (it should be 0.36 instead of 0.35).
>
> If the Zenodo authors are different on purpose, I'm not sure whether this is allowed by JOSS, so maybe the editor @AnjaliSandip can comment on that?

We have now made sure that the JOSS list of authors matchs the list of authors in the zenodo archive by adding Rick Lumpkin to the .zenodo.json file. His addition as an author will be triggered with the next release if @AnjaliSandip thinks that is appropriate?

> Other than that, in my opinion, the submission is in line with JOSS standards and is ready for publication. Clouddrift looks like a great tool to deal with Lagrangian data. I will advertise it with my former group at JHU that works extensively with Lagrangian analysis applied to physical oceanography. On this regard, you might find this JOSS paper interesting: [Seaduck](https://joss.theoj.org/papers/10.21105/joss.05967)

Thank you! We will definitely have a look at Seaduck.
