### 2024-06-24 #6742 @selipot
@malmans2 the title of the zenodo archive should now reflect the correct version, matching the release.
