### 2024-06-25 #6742 @rcaneill
Hi @selipot, thanks for addressing the comments, I have now checked all boxes of the checklist. Thanks to the updated notebooks, I have been able to tests the library on binder and everything works as expected!

I have one last comment that should not be hard to address: I struggled to understand how to produce a compatible xarray Dataset. Correct me if I’m wrong, but you only need a dataset with 1 dimension that represent the observation (e.g. "obs"), one dimension that represent each individual particle (e.g. "traj"), and one variable along "traj" that contains the rowsize.
If I understood this correctly, could you provide a minimal example with synthetic data in the documentation? This would be more or less the same example as the velocity in the paper, but written as xarray Dataset. I believe that it would help new beginners.

When this last issue is fixed, to my opinion, the publication will be ready for publication!
