### 2024-06-25 072a488
added ragged array example in usage.rst
