### 2024-06-26 #6742 @rcaneill
> @rcaneill I have a [PR](https://github.com/Cloud-Drift/clouddrift/pull/483) to add the example you requested. Would you be able to have a look?

@selipot this is perfect :)
