### 2024-06-26 ad08d4e
Doc edits (#483)
* reorder datasets.rst

* added ragged array example in usage.rst

* break example

* break example in two
