### 2024-06-26 #6742 @malmans2
> Thanks @selipot! @rcaneill @malmans2 is the paper ready for publication?

Yes!|
