### 2024-06-27 68a3cca
Improve/fix attrs and add dataset test
