### 2024-06-28 7ad3916
🔍 Update index.rst (#486)
* Update index.rst

Fix link that points to a fork repo

* fix type comparison
