+++ paper.bib
@@ -57,20 +57,20 @@ DOI = {10.5194/gmd-12-3571-2019}
   pages={2059--2077},
   year={2015},
   publisher={American Meteorological Society},
-  DOI = {https://doi.org/10.1175/BAMS-D-14-00110.1}
+  DOI = {10.1175/BAMS-D-14-00110.1}
 }
 
-@article{Maze:2020, 
-  doi = {10.21105/joss.02425}, 
-  url = {https://doi.org/10.21105/joss.02425}, 
-  year = {2020}, 
-  publisher = {The Open Journal}, 
-  volume = {5}, 
-  number = {53}, 
-  pages = {2425}, 
-  author = {Guillaume Maze and Kevin Balem}, 
-  title = {argopy: {A} {Python} library for {Argo} ocean data analysis}, 
-  journal = {Journal of Open Source Software} 
+@article{Maze:2020,
+  doi = {10.21105/joss.02425},
+  url = {https://doi.org/10.21105/joss.02425},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {53},
+  pages = {2425},
+  author = {Guillaume Maze and Kevin Balem},
+  title = {argopy: {A} {Python} library for {Argo} ocean data analysis},
+  journal = {Journal of Open Source Software}
 }
 
 @article{Lilly:2021,
@@ -89,7 +89,7 @@ DOI = {10.5194/gmd-12-3571-2019}
   number={1},
   pages={10--10},
   year={2017},
-  doi = {https://doi.org/10.5334/jors.148}
+  doi = {10.5334/jors.148}
 }
 
 @Article{Harris:2020,
@@ -161,13 +161,13 @@ DOI = {10.5194/gmd-12-3571-2019}
 }
 
 @article{Stern:2022,
-AUTHOR={Stern, Charles and Abernathey, Ryan and Hamman, Joseph and Wegener, Rachel and Lepore, Chiara and Harkins, Sean and Merose, Alexander},   
-TITLE={{Pangeo Forge: Crowdsourcing Analysis-Ready, Cloud Optimized Data Production}},      
-JOURNAL={Frontiers in Climate},      
-VOLUME={3},           
-YEAR={2022},      
-URL={https://www.frontiersin.org/articles/10.3389/fclim.2021.782909},       
-DOI={10.3389/fclim.2021.782909},      
-ISSN={2624-9553},   
+AUTHOR={Stern, Charles and Abernathey, Ryan and Hamman, Joseph and Wegener, Rachel and Lepore, Chiara and Harkins, Sean and Merose, Alexander},
+TITLE={{Pangeo Forge: Crowdsourcing Analysis-Ready, Cloud Optimized Data Production}},
+JOURNAL={Frontiers in Climate},
+VOLUME={3},
+YEAR={2022},
+URL={https://www.frontiersin.org/articles/10.3389/fclim.2021.782909},
+DOI={10.3389/fclim.2021.782909},
+ISSN={2624-9553},
 ABSTRACT={Pangeo Forge is a new community-driven platform that accelerates science by providing high-level recipe frameworks alongside cloud compute infrastructure for extracting data from provider archives, transforming it into analysis-ready, cloud-optimized (ARCO) data stores, and providing a human- and machine-readable catalog for browsing and loading. In abstracting the scientific domain logic of data recipes from cloud infrastructure concerns, Pangeo Forge aims to open a door for a broader community of scientists to participate in ARCO data production. A wholly open-source platform composed of multiple modular components, Pangeo Forge presents a foundation for the practice of reproducible, cloud-native, big-data ocean, weather, and climate science without relying on proprietary or cloud-vendor-specific tooling.}
 }
