### 2024-06-28 #6742 @AnjaliSandip
@selipot can you make a new release of the software being reviewed and deposit a new archive of the software with Zenodo/figshare?
