+++ pyproject.toml
@@ -4,7 +4,8 @@ build-backend = "hatchling.build"
 
 [project]
 name = "clouddrift"
-version = "0.37.0"
+version = "0.38.0"
+
 authors = [
   { name="Shane Elipot", email="selipot@miami.edu" },
   { name="Philippe Miron", email="philippemiron@gmail.com" },
