+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "clouddrift"
-version = "0.38.0"
+version = "0.38.1"
 
 authors = [
   { name="Shane Elipot", email="selipot@miami.edu" },
