+++ _sources/_autosummary/clouddrift.adapters.gdp1h.preprocess.rst
@@ -1,0 +1,6 @@
+clouddrift.adapters.gdp1h.preprocess
+====================================
+
+.. currentmodule:: clouddrift.adapters.gdp1h
+
+.. autofunction:: preprocess
