+++ _sources/_autosummary/clouddrift.signal.analytic_signal.rst
@@ -1,0 +1,6 @@
+clouddrift.signal.analytic\_signal
+==================================
+
+.. currentmodule:: clouddrift.signal
+
+.. autofunction:: analytic_signal
