+++ _sources/_autosummary/clouddrift.datasets.rst
@@ -1,0 +1,39 @@
+﻿clouddrift.datasets
+===================
+
+.. automodule:: clouddrift.datasets
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      andro
+      gdp1h
+      gdp6h
+      glad
+      hurdat2
+      mosaic
+      spotters
+      subsurface_floats
+      yomaha
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
