+++ _sources/_autosummary/clouddrift.adapters.rst
@@ -1,0 +1,42 @@
+﻿clouddrift.adapters
+===================
+
+.. automodule:: clouddrift.adapters
+   :members:
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :template: module.rst
+   :recursive:
+
+   clouddrift.adapters.andro
+   clouddrift.adapters.gdp
+   clouddrift.adapters.gdp1h
+   clouddrift.adapters.gdp6h
+   clouddrift.adapters.glad
+   clouddrift.adapters.hurdat2
+   clouddrift.adapters.mosaic
+   clouddrift.adapters.subsurface_floats
+   clouddrift.adapters.utils
+   clouddrift.adapters.yomaha
+
