+++ clouddrift/datasets.py
@@ -576,36 +576,36 @@ def andro(decode_times: bool = True) -> xr.Dataset:
     >>> from clouddrift.datasets import andro
     >>> ds = andro()
     >>> ds
-    <xarray.Dataset>
-    Dimensions:     (obs: 1360753, traj: 9996)
+    <xarray.Dataset> Size: 110MB
+    Dimensions:     (obs: 510630, traj: 3344)
     Coordinates:
-        time_d      (obs) datetime64[ns] ...
-        time_s      (obs) datetime64[ns] ...
-        time_lp     (obs) datetime64[ns] ...
-        time_lc     (obs) datetime64[ns] ...
-        id          (traj) int64 ...
+        time_d      (obs) datetime64[ns] 4MB ...
+        time_s      (obs) datetime64[ns] 4MB ...
+        time_lp     (obs) datetime64[ns] 4MB ...
+        time_lc     (obs) datetime64[ns] 4MB ...
+        id          (traj) float32 13kB ...
     Dimensions without coordinates: obs, traj
     Data variables: (12/33)
-        lon_d       (obs) float64 ...
-        lat_d       (obs) float64 ...
-        pres_d      (obs) float32 ...
-        temp_d      (obs) float32 ...
-        sal_d       (obs) float32 ...
-        ve_d        (obs) float32 ...
+        lon_d       (obs) float64 4MB ...
+        lat_d       (obs) float64 4MB ...
+        pres_d      (obs) float32 2MB ...
+        temp_d      (obs) float32 2MB ...
+        sal_d       (obs) float32 2MB ...
+        ve_d        (obs) float32 2MB ...
         ...          ...
-        lon_lc      (obs) float64 ...
-        lat_lc      (obs) float64 ...
-        surf_fix    (obs) int64 ...
-        cycle       (obs) int64 ...
-        profile_id  (obs) float32 ...
-        rowsize     (traj) int64 ...
+        lon_lc      (obs) float64 4MB ...
+        lat_lc      (obs) float64 4MB ...
+        surf_fix    (obs) float32 2MB ...
+        cycle       (obs) float32 2MB ...
+        profile_id  (obs) float32 2MB ...
+        rowsize     (traj) int64 27kB ...
     Attributes:
-        title:           ANDRO: An Argo-based deep displacement dataset
-        history:         2022-03-04
-        date_created:    2023-12-08T00:52:00.937120
+        title:           ANDRO: An Argo-based deep displacement dataset (Quality ...
+        history:         Dataset updated on 2024-03-25
+        date_created:    2024-07-02T12:11:50.643492
         publisher_name:  SEANOE (SEA scieNtific Open data Edition)
         publisher_url:   https://www.seanoe.org/data/00360/47077/
-        license:         freely available
+        license:         Creative Commons Attribution 4.0 International License (...
 
     Reference
     ---------
