### 2024-07-08 #6742 @selipot
Thank you @AnjaliSandip for handling the editorial process and @malmans2 and @rcaneill for your thorough reviews. We are just waiting for [this pull request](https://github.com/openjournals/joss-papers/pull/5588)  to be merged for the official publication, is this correct?
