### 2024-07-13 a60edbc
Enhance andro tests
