+++ tests/datasets_tests.py
@@ -53,8 +53,12 @@ class datasets_tests(testutils.DisableProgressTestCase):
             self.assertTrue(len(ds.variables) > 0)
             self.assertTrue(len(ds["lon_d"]) > 0)
 
-            self.assertTrue(len(ds.lat_d[np.logical_or(ds.lat_d > 90, ds.lat_d < -90)]) == 0)
-            self.assertTrue(len(ds.lat_d[np.logical_or(ds.lon_d > 180, ds.lon_d < -180)]) == 0)
+            self.assertTrue(
+                len(ds.lat_d[np.logical_or(ds.lat_d > 90, ds.lat_d < -90)]) == 0
+            )
+            self.assertTrue(
+                len(ds.lat_d[np.logical_or(ds.lon_d > 180, ds.lon_d < -180)]) == 0
+            )
 
     def test_yomaha_opens(self):
         with datasets.yomaha() as ds:
