### 2024-07-13 dddc803
🐛 fix the andro datasets (#489)
* fix url

* put back decode_times default value to True

* update docstring

* Enhance andro tests

* format

* bump version

---------

Co-authored-by: Kevin Santana <kevinsantana11@gmail.com>
