### 2024-07-15 01e2d4a
replace type comparison with isinstance
