### 2024-07-16 079ae36
replace type comparison with isinstance (#491)
