### 2024-07-16 #6742 @kthyng
Sorry for my delay! I was at the SciPy conference last week. On the plus side, I met your co-author @kevinsantana11 there!

Moving forward, I'll take over now as Track Associate Editor in Chief to do some final submission editing checks. After these checks are complete, I will publish your submission!
