### 2024-07-16 #6742 @kthyng
@selipot Can you update the Zenodo metadata so the title of v0.38.1 (the JOSS paper version) matches your JOSS paper? This is a request not a requirement, to make a cohesive archive across the products.
