### 2024-07-16 #6742 @philippemiron
Could we set the version here to v0.39 instead?
