### 2024-07-17 #6742 @kevinsantana11
Unsure if I could run this but I think we just need to include a comment such as:

`at_editorialbot set v0.39.0 as version` (at_ -> @) so the version referenced in this PR matches up with the latest version in zenodo.
