### 2024-07-17 #6742 @kthyng
`v0.38.1` is the one in Zenodo noted as the JOSS version, plus I assume the reviewers didn't review the changes since then. What is your reasoning for updating the version?
