### 2024-07-17 #6742 @philippemiron
Actually, the reviewer looked at `v0.37.0`, and we were asked to publish a _new_ version after the review was completed (https://github.com/openjournals/joss-reviews/issues/6742#issuecomment-2197079307).

After the review was completed, I thought we could keep going with the normal release schedule of the library. Release `v0.39.0` is minimal and has only two changes (https://github.com/Cloud-Drift/clouddrift/releases/tag/v0.39.0)
- 🔎 update gdp6h docstring
- 🐛 fix the andro datasets

The Zendo records https://zenodo.org/records/12737923 get updated automatically with the pipeline. If required, we could use this record, https://zenodo.org/records/12583739, which has the `v0.38.1` tag.
