### 2024-07-17 #6742 @kthyng
Yes, it is typical to have a release after the review process to capture any changes in your software that might have come about due to the review itself, and yes, you can continue to change your codebase over time (separate from the review, I know this final publication process took some extra time). However, we want to capture the release that is associated with the review itself. I understand this to be `v0.38.1`, is that right?
