### 2024-07-17 #6742 @kevinsantana11
@kthyng that is correct. If we want to reference the library version release approved by the reviewers then we would need to change the zenodo archive DOI to the v0.38.1, which is: https://doi.org/10.5281/zenodo.12583739. The problem with the current DOI, is that it references the latest version which is currently v0.39.0 vs the one reviewed (v.0.38.1) and approved.
