### 2024-07-20 dc13f9a
chore: update GDP_VERSION to 2.01.1 and fix location_type attribute comparison
