+++ clouddrift/adapters/gdp1h.py
@@ -20,7 +20,7 @@ import clouddrift.adapters.gdp as gdp
 from clouddrift.adapters.utils import download_with_progress
 from clouddrift.raggedarray import RaggedArray
 
-GDP_VERSION = "2.01"
+GDP_VERSION = "2.01.1"
 
 
 GDP_DATA_URL = "https://www.aoml.noaa.gov/ftp/pub/phod/buoydata/hourly_product/v2.01"
@@ -212,7 +212,7 @@ def preprocess(index: int, **kwargs) -> xr.Dataset:
     # convert attributes to variable
     ds["location_type"] = (
         ("traj"),
-        [False if ds.get("location_type") == "Argos" else True],
+        [False if ds.attrs.get("location_type") == "Argos" else True],
     )  # 0 for Argos, 1 for GPS
     ds["DeployingShip"] = (("traj"), gdp.cut_str(ds.DeployingShip, 20))
     ds["DeploymentStatus"] = (
