### 2024-07-21 c525b20
🐛 Include connection errors as retryable exceptions
