### 2024-07-21 35b44ee
Update test to ensure location_type is correctly derived from netcdf file attributes
