### 2024-07-21 a8eba57
add more exception
