### 2024-07-21 04ff774
feat: Update random number generator in gdp1h and gdp6h adapters
The code changes update the random number generator in the `gdp1h.py` and `gdp6h.py` adapters. The previous usage of `np.random.RandomState` has been replaced with `np.random.Generator(np.random.MT19937)`. This change ensures consistent and reproducible random number generation.
