### 2024-07-21 0537850
chore: Update type comparison in ragged.py
The code changes in ragged.py update the type comparison for the `tolerance` variable. Instead of using the `type()` function, it now uses the `isinstance()` function to check if `tolerance` is an instance of `pd.Timedelta`. This change improves the accuracy and reliability of the comparison.

Note: This commit message follows the established convention of using the "chore" type for non-functional changes that don't affect the code's behavior.
