### 2024-07-21 cced3e7
Close buffer from top level logic
