+++ tests/adapters/utils_tests.py
@@ -81,8 +81,9 @@ class utils_tests(unittest.TestCase):
                 ),
             ]
         ) as _:
+            buffer = Mock()
             utils._download_with_progress(
-                "some.url.com", "./some/path/existing-file.nc", 0, False
+                "some.url.com", buffer, 0, False
             )
             self.requests_mock.get.assert_called()
 
@@ -166,8 +167,9 @@ class utils_tests(unittest.TestCase):
                 patch("clouddrift.adapters.utils.requests", self.requests_mock),
             ]
         ) as mock:
+            buffer = Mock()
             utils._download_with_progress(
-                "some.url.com", "./some/path/existing-file.nc", 0, True
+                "some.url.com", buffer, 0, True
             )
             mock[0].assert_called()
             self.bar_mock.update.assert_called()
@@ -185,8 +187,9 @@ class utils_tests(unittest.TestCase):
                 patch("clouddrift.adapters.utils.requests", self.requests_mock),
             ]
         ) as mock:
+            buffer = Mock()
             utils._download_with_progress(
-                "some.url.com", "./some/path/existing-file.nc", 0, False
+                "some.url.com", buffer, 0, False
             )
             mock[0].assert_not_called()
             self.bar_mock.update.assert_not_called()
@@ -235,7 +238,7 @@ class utils_tests(unittest.TestCase):
             )
             assert tpe_mock.submit.call_count == len(mocked_futures)
             assert self.bar_mock.update.call_count == 2
-            assert os_mock.remove.call_count == len(mocked_futures) - 1
+            assert os_mock.remove.call_count == len(mocked_futures)
             tpe_mock.shutdown.assert_called_once()
             [fut_mock.cancel.assert_called_once() for fut_mock in mocked_futures[:-1]]
             mocked_futures[-1].cancel.assert_not_called()
