+++ clouddrift/adapters/gdp1h.py
@@ -139,7 +139,7 @@ def preprocess(index: int, **kwargs) -> xr.Dataset:
         fp,
         decode_times=False,
         decode_coords=False,
-        engine = "netcdf4",
+        engine="netcdf4",
     )
 
     # parse the date with custom function
