### 2024-07-21 0696210
decompress entire buffer
