+++ clouddrift/adapters/gdp.py
@@ -11,7 +11,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from clouddrift.adapters.utils import download_with_progress
+from clouddrift.adapters.utils import download_with_progress, standard_retry_protocol
 from clouddrift.raggedarray import DimNames
 
 GDP_DIMS: dict[str, DimNames] = {"traj": "rows", "obs": "obs"}
@@ -106,9 +106,9 @@ def parse_directory_file(filename: str) -> pd.DataFrame:
         List of drifters from a single directory file as a pandas DataFrame.
     """
     GDP_DIRECTORY_FILE_URL = "https://www.aoml.noaa.gov/ftp/pub/phod/buoydata/"
-    df = pd.read_csv(
+    df = standard_retry_protocol(lambda: pd.read_csv(
         os.path.join(GDP_DIRECTORY_FILE_URL, filename), delimiter=r"\s+", header=None
-    )
+    ))()
 
     # Combine the date and time columns to easily parse dates below.
     df[4] += " " + df[5]
