+++ CONTRIBUTING.md
@@ -63,7 +63,7 @@ Common emojis to use are as follow:
 🐛 Bugfix
 🧹 Chore / refactoring / migration / dependency deprecation related
 ⚡ Performance/memory improvements
-🔍 Documentation, refactoring
+🔍 Documentation
 🔧 Tooling/Build scripts/CI (other non-application changes)
 ```
 
