+++ clouddrift/adapters/gdp1h.py
@@ -215,85 +215,96 @@ def preprocess(index: int, **kwargs) -> xr.Dataset:
         ("traj"),
         [False if ds.attrs.get("location_type") == "Argos" else True],
     )  # 0 for Argos, 1 for GPS
-    ds["DeployingShip"] = (("traj"), gdp.cut_str(ds.DeployingShip, 20))
+    ds["DeployingShip"] = (("traj"), gdp.cut_str(ds.attrs.get("DeployingShip", ""), 20))
     ds["DeploymentStatus"] = (
         ("traj"),
-        gdp.cut_str(ds.DeploymentStatus, 20),
+        gdp.cut_str(ds.attrs.get("DeploymentStatus", ""), 20),
     )
     ds["BuoyTypeManufacturer"] = (
         ("traj"),
-        gdp.cut_str(ds.BuoyTypeManufacturer, 20),
+        gdp.cut_str(ds.attrs.get("BuoyTypeManufacturer", ""), 20),
     )
     ds["BuoyTypeSensorArray"] = (
         ("traj"),
-        gdp.cut_str(ds.BuoyTypeSensorArray, 20),
+        gdp.cut_str(ds.attrs.get("BuoyTypeSensorArray", ""), 20),
     )
     ds["CurrentProgram"] = (
         ("traj"),
-        np.array([gdp.str_to_float(ds.CurrentProgram, -1)], dtype=np.int32),
+        np.array(
+            [gdp.str_to_float(ds.attrs.get("CurrentProgram", ""), -1)], dtype=np.int32
+        ),
     )
     ds["PurchaserFunding"] = (
         ("traj"),
-        gdp.cut_str(ds.PurchaserFunding, 20),
+        gdp.cut_str(ds.attrs.get("PurchaserFunding", ""), 20),
     )
-    ds["SensorUpgrade"] = (("traj"), gdp.cut_str(ds.SensorUpgrade, 20))
-    ds["Transmissions"] = (("traj"), gdp.cut_str(ds.Transmissions, 20))
+    ds["SensorUpgrade"] = (("traj"), gdp.cut_str(ds.attrs.get(".SensorUpgrade"), 20))
+    ds["Transmissions"] = (("traj"), gdp.cut_str(ds.attrs.get("Transmissions"), 20))
     ds["DeployingCountry"] = (
         ("traj"),
-        gdp.cut_str(ds.DeployingCountry, 20),
+        gdp.cut_str(ds.attrs.get("DeployingCountry", ""), 20),
     )
     ds["DeploymentComments"] = (
         ("traj"),
-        np.array([gdp.str_to_float(ds.ManufactureYear, -1)], dtype=np.int16),
+        np.array(
+            [gdp.str_to_float(ds.attrs.get("ManufactureYear", ""), -1)], dtype=np.int16
+        ),
     )
     ds["ManufactureMonth"] = (
         ("traj"),
-        np.array([gdp.str_to_float(ds.ManufactureMonth, -1)], dtype=np.int16),
+        np.array(
+            [gdp.str_to_float(ds.attrs.get("ManufactureMonth", ""), -1)], dtype=np.int16
+        ),
     )
     ds["ManufactureSensorType"] = (
         ("traj"),
-        gdp.cut_str(ds.ManufactureSensorType, 20),
+        gdp.cut_str(ds.attrs.get("ManufactureSensorType", ""), 20),
     )
     ds["ManufactureVoltage"] = (
         ("traj"),
-        np.array([gdp.str_to_float(ds.ManufactureVoltage[:-2], -1)], dtype=np.int16),
+        np.array(
[…]
