### 2024-07-21 682bf56
generalize print message
