+++ clouddrift/adapters/utils.py
@@ -23,10 +23,8 @@ _DISABLE_SHOW_PROGRESS = False  # purely to de-noise our test suite output, shou
 
 def _before_call(rcs: RetryCallState):
     if rcs.attempt_number > 1:
-        src = rcs.args[0]
-        dst = "io-buffer" if isinstance(rcs.args[1], BufferedIOBase) else rcs.args[1]
         _logger.warn(
-            f"retrying download request for (dst, src): {(src, dst)}, attempt: {rcs.attempt_number}"
+            f"retrying request for (fn, args): {(rcs.fn.__name__, rcs.args)}, attempt: {rcs.attempt_number}"
         )
 
 
