### 2024-07-22 562fbcb
read should be part of the retryable logic
