### 2024-07-22 b4da269
non buggy decompression logic
