### 2024-07-22 4c52bd2
🐛 fix locationtype bug in dataset (#494)
* fixes #492

* chore: update GDP_VERSION to 2.01.1 and fix location_type attribute comparison

* Update test to ensure location_type is correctly derived from netcdf file attributes

* get values from attrs property of an xr dataset vs directly from object.

---------

Co-authored-by: Shane Elipot <selipot@miami.edu>
Co-authored-by: Philippe Miron <philippemiron@gmail.com>
