### 2024-07-22 #6742 @kevinsantana11
Hi @kthyng following up here I believe we need to run:

so that this review is associated to the the clouddrift release @ v0.38.1 which has been approved by the reviewers. If there's something needed on our end, let us know :smile:
