### 2024-07-22 #6742 @kthyng
@selipot @philippemiron @kevinsantana11 Great. We also prefer to have the title and author list in the metadata for that JOSS Zenodo archive match your JOSS paper — can you update that? It looks like the author list is in the correct order, but are your first and last names backward?
