+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "clouddrift"
-version = "0.39.0"
+version = "0.40.0"
 
 authors = [
   { name="Shane Elipot", email="selipot@miami.edu" },
