+++ _sources/_autosummary/clouddrift.adapters.gdp6h.download.rst
@@ -1,0 +1,6 @@
+clouddrift.adapters.gdp6h.download
+==================================
+
+.. currentmodule:: clouddrift.adapters.gdp6h
+
+.. autofunction:: download
