+++ _sources/_autosummary/clouddrift.adapters.gdp.rst
@@ -1,0 +1,41 @@
+clouddrift.adapters.gdp
+=======================
+
+.. automodule:: clouddrift.adapters.gdp
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      cast_float64_variables_to_float32
+      cut_str
+      decode_date
+      drogue_presence
+      fetch_netcdf
+      fill_values
+      get_gdp_metadata
+      order_by_date
+      parse_directory_file
+      rowsize
+      str_to_float
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
