+++ _sources/_autosummary/clouddrift.adapters.mosaic.get_repository_metadata.rst
@@ -1,0 +1,6 @@
+clouddrift.adapters.mosaic.get\_repository\_metadata
+====================================================
+
+.. currentmodule:: clouddrift.adapters.mosaic
+
+.. autofunction:: get_repository_metadata
