+++ _sources/_autosummary/clouddrift.adapters.rst
@@ -1,0 +1,42 @@
+﻿clouddrift.adapters
+===================
+
+.. automodule:: clouddrift.adapters
+   :members:
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :template: module.rst
+   :recursive:
+
+   andro
+   gdp
+   gdp1h
+   gdp6h
+   glad
+   hurdat2
+   mosaic
+   subsurface_floats
+   utils
+   yomaha
+
