+++ _sources/_autosummary/clouddrift.adapters.gdp6h.to_raggedarray.rst
@@ -1,0 +1,6 @@
+clouddrift.adapters.gdp6h.to\_raggedarray
+=========================================
+
+.. currentmodule:: clouddrift.adapters.gdp6h
+
+.. autofunction:: to_raggedarray
