+++ _sources/_autosummary/clouddrift.adapters.gdp1h.rst
@@ -1,0 +1,33 @@
+clouddrift.adapters.gdp1h
+=========================
+
+.. automodule:: clouddrift.adapters.gdp1h
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      download
+      preprocess
+      to_raggedarray
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
