+++ _sources/_autosummary/clouddrift.adapters.subsurface_floats.rst
@@ -1,0 +1,32 @@
+clouddrift.adapters.subsurface\_floats
+======================================
+
+.. automodule:: clouddrift.adapters.subsurface_floats
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      download
+      to_xarray
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
