+++ _sources/_autosummary/clouddrift.adapters.gdp6h.rst
@@ -1,0 +1,33 @@
+clouddrift.adapters.gdp6h
+=========================
+
+.. automodule:: clouddrift.adapters.gdp6h
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      download
+      preprocess
+      to_raggedarray
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
