+++ _sources/_autosummary/clouddrift.adapters.gdp.fill_values.rst
@@ -1,0 +1,6 @@
+clouddrift.adapters.gdp.fill\_values
+====================================
+
+.. currentmodule:: clouddrift.adapters.gdp
+
+.. autofunction:: fill_values
