+++ _sources/_autosummary/clouddrift.ragged.rst
@@ -1,0 +1,39 @@
+﻿clouddrift.ragged
+=================
+
+.. automodule:: clouddrift.ragged
+   :members:
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+   
+      apply_ragged
+      chunk
+      prune
+      ragged_to_regular
+      regular_to_ragged
+      rowsize_to_index
+      segment
+      subset
+      unpack
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
