### 2024-07-23 #6742 @kthyng
@kevinsantana11 You can edit the metadata to change the author list and title of previous versions. Here is a [link](https://help.zenodo.org/docs/deposit/manage-versions/#:~:text=Edit%20metadata%20of%20previous%20version,-1&text=Find%20the%20previous%20version%20you,(see%20edit%20published%20records)) with a little help.
