### 2024-07-23 #6742 @kevinsantana11
@kthyng the metadata has been updated for all releases including v0.38.1 which was reviewed by the committee
