### 2024-07-23 #6742 @kthyng
Great. As I said before, we also prefer to have the title in the metadata for that JOSS Zenodo archive match your JOSS paper. Can you update that or do you prefer not to? This is not a requirement but a way to have a cohesive archive for that particular version.
