### 2024-07-23 #6742 @selipot
i am trying to understand here: You want the title of the zenodo archive to be "Clouddrift: a Python package to accelerate the use of Lagrangian data for atmospheric, oceanic, and climate sciences" instead of "Cloud-Drift/clouddrift: v0.38.1"?
