### 2024-07-23 #6742 @selipot
If that's ok I'd like to pass on this request. I would like to have different titles for the zenodo archive of the github repo on one hand and the JOSS paper on the other hand.
