### 2024-07-23 #6742 @kthyng
Congratulations on your new publication @selipot, @philippemiron, @kevinsantana11! Many thanks to @AnjaliSandip and to reviewers @rcaneill and @malmans2 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@selipot, @kevinsantana11 Please sign up at https://reviewers.joss.theoj.org/ if you'd like to review with JOSS in the future!
